### 2021-11-12 #3908 @kyleniemeyer
@amoradinejad I also noticed that the README describes the software documentation as still under development. For us to proceed with the review here, we expect sufficient documentation, as described in our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation).

I'm also not seeing anything in the software repository that looks like tests. Are there any tests or example cases built in that would enable reviewers to evaluate that the software is working correctly?
