### 2021-11-12 #3908 @arfon
> For some reason @whedon is not reporting the repository/code statistics completely, I am not sure why. Also, the paper does have a "Statement of Need" section, it is just part of a "Scientific Context and Statement of Need" section title.

@kyleniemeyer – I ran `cloc` locally:

```
Language                     files          blank        comment           code
C/C++ Header                    41           1561           1359        1578221
HTML                           133            894            240          19280
TeX                             32           1054            323           8855
C                               14           1556           1969           4658
JavaScript                     116            159            162           2837
CSS                              4            365             51           1833
XML                              1              6              8            226
SVG                              3              1              1            129
Markdown                         3             53              0             94
make                             2              7              0             49
YAML                             1              1              4             18
SUM:                           350           5657           4117        1616200
```
