+++ Paper/paper.bib
@@ -106,7 +106,10 @@
     eprint = "2103.01971",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
-    month = "3",
+    doi = "10.1088/1475-7516/2021/05/067",
+    journal = "JCAP",
+    volume = "05",
+    pages = "067",
     year = "2021"
 }
 
