### 2021-11-13 #3908 @amoradinejad
I resolved the reference issue, listed below.

MISSING DOIs

- 10.1088/1475-7516/2021/05/067 may be a valid DOI for title: Astrophysics & Cosmology from Line Intensity Mapping vs Galaxy Surveys

INVALID DOIs

Best,
Azadeh Moradinezhad

> On 12 Nov 2021, at 19:31, Kyle Niemeyer ***@***.***> wrote:
>
>
> For some reason @whedon <https://github.com/whedon> is not reporting the repository/code statistics completely, I am not sure why. Also, the paper does have a "Statement of Need" section, it is just part of a "Scientific Context and Statement of Need" section title.
>
> @amoradinejad <https://github.com/amoradinejad> while we are getting started, can you resolve the reference issues pointed out above?
>
