### 2021-11-13 c4b2514
two fo the functions in wnw_split module commented
