+++ docs/latex/df/dcb/header_8h.tex
@@ -39,14 +39,16 @@
 {\ttfamily \#include \char`\"{}ps\+\_\+line\+\_\+pt.\+h\char`\"{}}\newline
 {\ttfamily \#include \char`\"{}ps\+\_\+line\+\_\+hm.\+h\char`\"{}}\newline
 {\ttfamily \#include \char`\"{}cubature.\+h\char`\"{}}\newline
-Include dependency graph for header.\+h\+:\nopagebreak
+Include dependency graph for header.\+h\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{db/d0a/header_8h__incl}
 \end{center}
 \end{figure}
-This graph shows which files directly or indirectly include this file\+:\nopagebreak
+This graph shows which files directly or indirectly include this file\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -498,14 +500,16 @@ In units of km/s
 
 omega\+\_\+b = Omega\+\_\+b h$^\wedge$2;
 
-3.\+0665Here is the call graph for this function\+:\nopagebreak
+3.\+0665Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{df/dcb/header_8h_a25a40b6614565f755233080a384c35f1_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
