+++ docs/latex/refman.tex
@@ -15,6 +15,7 @@
   % Packages required by doxygen
   \usepackage{fixltx2e} % for \textsubscript
   \usepackage{doxygen}
+  \usepackage{doxygen}
   \usepackage{graphicx}
   \usepackage[utf8]{inputenc}
   \usepackage{makeidx}
