+++ docs/latex/d0/d29/main_8c.tex
@@ -6,8 +6,7 @@ Documented main module, including functions to initilize and cleanup the cosmolo
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for main.\+c\+:
-\nopagebreak
+Include dependency graph for main.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -51,16 +50,14 @@ Free the memory allocated to cosmology structure and SFR interpolator.
 \begin{DoxyReturn}{Returns}
 void 
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d0/d29/main_8c_a798f0df7ac8f26aae66b7364d7f4e6f0_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -101,8 +98,7 @@ omega\+\_\+b = Omega\+\_\+b h$^\wedge$2;
 \includegraphics[width=350pt]{d0/d29/main_8c_a25a40b6614565f755233080a384c35f1_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
