+++ docs/latex/files.tex
@@ -11,6 +11,5 @@ Here is a list of all files with brief descriptions\+:\begin{DoxyCompactList}
 \item\contentsline{section}{\mbox{\hyperlink{ps__line__pt_8c}{ps\+\_\+line\+\_\+pt.\+c}} \\*Documented computation of Poisson shot noise and tree-\/level line power spectrum in real and redshift-\/space }{\pageref{ps__line__pt_8c}}{}
 \item\contentsline{section}{\mbox{\hyperlink{survey__specs_8c}{survey\+\_\+specs.\+c}} \\*Documented computation of some survey-\/related functions }{\pageref{survey__specs_8c}}{}
 \item\contentsline{section}{\mbox{\hyperlink{utilities_8c}{utilities.\+c}} \\*Documented basic utility functions used by other modules of the code }{\pageref{utilities_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{wnw__split_8c}{wnw\+\_\+split.\+c}} \\*Documented wiggle-\/nowiggle split based on 3d Gaussian filter in linear k, and using the Eisentein-\/\+Hu wiggle-\/no wiggle template ~\newline
- }{\pageref{wnw__split_8c}}{}
+\item\contentsline{section}{\mbox{\hyperlink{wnw__split_8c}{wnw\+\_\+split.\+c}} \\*Documented wiggle-\/nowiggle split based on 3d Gaussian filter in linear k, and using the Eisentein-\/\+Hu wiggle-\/no wiggle template ar\+Xiv\+:astro-\/ph/9709112 }{\pageref{wnw__split_8c}}{}
 \end{DoxyCompactList}
