+++ docs/latex/d8/d2a/ps__halo__1loop_8c.tex
@@ -6,8 +6,7 @@ Documented real-\/space, direct integration computation of 1loop contributions o
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+halo\+\_\+1loop.\+c\+:
-\nopagebreak
+Include dependency graph for ps\+\_\+halo\+\_\+1loop.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -107,8 +106,7 @@ Here is the call graph for this function\+:
 \includegraphics[width=350pt]{d8/d2a/ps__halo__1loop_8c_a06812b0e40a5e228647283536b82fa5a_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -151,8 +149,7 @@ Here is the call graph for this function\+:
 \includegraphics[width=350pt]{d8/d2a/ps__halo__1loop_8c_a05a2d154a1db42273c26d91b1145b892_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -165,8 +162,7 @@ Here is the caller graph for this function\+:
 \doxyparagraph{\texorpdfstring{F2()}{F2()}}
 {\footnotesize\ttfamily double F2 (\begin{DoxyParamCaption}\item[{double}]{k1,  }\item[{double}]{k2,  }\item[{double}]{mu }\end{DoxyParamCaption})}
 
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -179,8 +175,7 @@ Here is the caller graph for this function\+:
 \doxyparagraph{\texorpdfstring{F2\_s()}{F2\_s()}}
 {\footnotesize\ttfamily double F2\+\_\+s (\begin{DoxyParamCaption}\item[{double}]{k1,  }\item[{double}]{k2,  }\item[{double}]{mu }\end{DoxyParamCaption})}
 
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -193,8 +188,7 @@ Here is the caller graph for this function\+:
 \doxyparagraph{\texorpdfstring{F3\_s()}{F3\_s()}}
 {\footnotesize\ttfamily double F3\+\_\+s (\begin{DoxyParamCaption}\item[{double}]{k,  }\item[{double}]{q,  }\item[{double}]{mu }\end{DoxyParamCaption})}
 
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -217,8 +211,7 @@ The integrands below correspond to the follwing bias combinaions\+:Here is the c
 \includegraphics[width=350pt]{d8/d2a/ps__halo__1loop_8c_a9d475dd2ab5b0120a8c78ba15080494b_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
[…]
