+++ docs/latex/dd/d24/survey__specs_8c.tex
@@ -6,8 +6,7 @@ Documented computation of some survey-\/related functions.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for survey\+\_\+specs.\+c\+:
-\nopagebreak
+Include dependency graph for survey\+\_\+specs.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -67,16 +66,14 @@ We follow Giannantonio et al. to for determining kmax, and use gsl Brent solver 
 \begin{DoxyReturn}{Returns}
 kmax 
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{dd/d24/survey__specs_8c_a1eb59c068328f2b2a2cad533600c0d78_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -122,8 +119,7 @@ Compute the size of fundumental mode corresponding to the comoving volume enclos
 \begin{DoxyReturn}{Returns}
 kmin 
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -152,16 +148,14 @@ Compute the comoving volume of a survey covering redshift up to z.
 \begin{DoxyReturn}{Returns}
 the comoving z-\/shell volume 
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=323pt]{dd/d24/survey__specs_8c_a91c4e0c62bcd3e179536fea62e8d53b5_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
