### 2021-11-13 #3908 @amoradinejad
Hi,

I address the two questions in-line below.

> On 12 Nov 2021, at 19:35, Kyle Niemeyer ***@***.***> wrote:
>
>
> @amoradinejad <https://github.com/amoradinejad> I also noticed that the README describes the software documentation as still under development. For us to proceed with the review here, we expect sufficient documentation, as described in our review criteria <https://joss.readthedocs.io/en/latest/review_criteria.html#documentation>.
>
>

Documentation is complete, since only two functions were missing comments. I updated those comments. The full documentation was generated by Oxygen, and is found on the url https://amoradinejad.github.io/limHaloPT/html/index.html <https://amoradinejad.github.io/limHaloPT/html/index.html> .

> I'm also not seeing anything in the software repository that looks like tests. Are there any tests or example cases built in that would enable reviewers to evaluate that the software is working correctly?
>
There is no separate test file. However, the example tests for function calls for some of the main functions , which compute the line signal are given in main.c module.

Please let me know if further information was needed.

Best,
Azadeh Moradinezhad
