### 2021-11-13 #3908 @amoradinejad
Hi,

Since my scientific publications are cited with my full last name, Moradinezhad Dizgah, may I ask yo have the citation label (footnote on page 1) to “Moradinezhad Dizgah et al” instead of “Dizgah et al” ?

Thank you,
Azadeh Moradinezhad

> On 13 Nov 2021, at 20:10, whedon ***@***.***> wrote:
>
>
> 👉📄 Download article proof <https://raw.githubusercontent.com/openjournals/joss-papers/joss.03908/joss.03908/10.21105.joss.03908.pdf> 📄 View article proof on GitHub <https://github.com/openjournals/joss-papers/blob/joss.03908/joss.03908/10.21105.joss.03908.pdf> 📄 👈
>
