### 2021-11-15 #3908 @kyleniemeyer
> Hi, Since my scientific publications are cited with my full last name, Moradinezhad Dizgah, may I ask yo have the citation label (footnote on page 1) to “Moradinezhad Dizgah et al” instead of “Dizgah et al” ? Thank you, Azadeh Moradinezhad
> […](#)

This may be something that @arfon has to do manually when we get to the publication stage.
