### 2021-11-16 #3908 @arfon
> This may be something that @arfon has to do manually when we get to the publication stage.

Yes, please mention this at the end of the review process @amoradinejad – we have to manually fix this when the paper is ready to be published.
