### 2021-12-22 #3908 @eloisabentivegna
Further update: @ziotom78 has accepted to review the paper (thanks!), so long as a second reviewer with deeper expertise in line intensity mapping (or cosmological power spectrum analysis in general) addresses these aspects. @sgeorge91, @changhoonhahn, @alaingdl, can any of you help here?
