### 2022-01-20 79571ab
edits to some comments, updated version of class to v3.1, which means cosmology.c module is edited to be compatible with functions of the new version of class
