+++ Source/IR_res.c
@@ -26,7 +26,6 @@
  * @param k            Input: wavenumber in unit of 1/Mpc. 
  * @param z            Input: redshift
  * @param SPLIT        Input: switch to set the method of wiggle-nowiggle split
- * 
  * @return value of leading IR-ressumed power spectrum           
  */
 double pm_IR_LO(struct Cosmology *Cx, double k, double z, long SPLIT)
@@ -37,7 +36,7 @@ double pm_IR_LO(struct Cosmology *Cx, double k, double z, long SPLIT)
           sig2_LO = IR_Sigma2(Cx,z, kf0, SPLIT);
     }
     double p_nowiggle = pm_nowiggle(Cx, k, z, kf0, 0, SPLIT);
-    double p_wiggle   = Pk_dlnPk(Cx, k, z, LPOWER) - p_nowiggle;
+    double p_wiggle   = PS(Cx, k, z) - p_nowiggle;
     double sup        = exp(-k * k * sig2_LO);
     double f          = p_nowiggle + sup * p_wiggle;
 
@@ -52,21 +51,20 @@ double pm_IR_LO(struct Cosmology *Cx, double k, double z, long SPLIT)
  * @param k            Input: wavenumber in unit of 1/Mpc. 
  * @param z            Input: redshift
  * @param SPLIT        Input: switch to set the method of wiggle-nowiggle split
- * 
  * @return value of NL IR-ressumed power spectrum           
  */
 double pm_IR_NLO(struct Cosmology *Cx, double k,  double z, long SPLIT)
 {
     extern struct globals gb;
     static double sig2_NLO = - 1.;
-    double kf0 = 1.e-4;
+    double k0 = 1.e-4;
 
     if(sig2_NLO == - 1.){
-      sig2_NLO = IR_Sigma2(Cx,z, kf0, SPLIT);
+      sig2_NLO = IR_Sigma2(Cx,z, k0, SPLIT);
     }
 
-    double p_nowiggle = pm_nowiggle(Cx, k, z, kf0, 0, SPLIT);
-    double p_wiggle   = Pk_dlnPk(Cx, k, z, LPOWER) - p_nowiggle;
+    double p_nowiggle = pm_nowiggle(Cx, k, z, k0, 0, SPLIT);
+    double p_wiggle   = PS(Cx, k, z) - p_nowiggle;
     double sup        = exp(-k * k * sig2_NLO);
 
     double *pm_loops  =  make_1Darray(2);
@@ -89,7 +87,6 @@ double pm_IR_NLO(struct Cosmology *Cx, double k,  double z, long SPLIT)
  * 
  * @param x            Input: integration variable, k-values
  * @param par          Input: integration parameters
- * 
  * @return integrand to be used in IR_sigma2() function         
  */
 double IR_Sigma2_integrand(double x, void *par)
@@ -102,12 +99,12 @@ double IR_Sigma2_integrand(double x, void *par)
     struct Cosmology *Cx = pij.p1;
     double bao_scale     = pij.p4;
     double z             = pij.p5;
-    double kf0           = pij.p6; 
+    double k0           = pij.p6; 
     long   SPLIT         = pij.p13;
 
     double k_osc = 1./bao_scale;  /// BAO_scale = 110. Mpc/h.
 
-    result = 1./(6.*M_PI*M_PI)*pm_nowiggle(Cx, x, z, kf0, 0, SPLIT)* (1. - gsl_sf_bessel_j0(x/k_osc) + 2. * gsl_sf_bessel_j2(x/k_osc));;
[…]
