+++ docs/latex/d2/dbd/struct_cosmology.tex
@@ -11,11 +11,12 @@ Structure that holds varioud quantities that need to be evaluated for a given ch
 
 
 
-Collaboration diagram for Cosmology\+:\nopagebreak
+Collaboration diagram for Cosmology\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=262pt]{d3/df4/struct_cosmology__coll__graph}
+\includegraphics[width=336pt]{d3/df4/struct_cosmology__coll__graph}
 \end{center}
 \end{figure}
 \doxysubsubsection*{Data Fields}
@@ -25,11 +26,13 @@ struct \mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_\+\+Cosmolog
 \item 
 struct \mbox{\hyperlink{struct_line}{Line}} $\ast$$\ast$ \mbox{\hyperlink{struct_cosmology_a0f59f743170aa24dbaeebd4058ac226a}{Lines}}
 \item 
+struct \mbox{\hyperlink{struct_p_s__xtr}{PS\+\_\+xtr}} $\ast$ \mbox{\hyperlink{struct_cosmology_a5dee8f06ef55f61fcf2ad920f4f80e65}{PS\+\_\+xtrapol}}
+\item 
 int \mbox{\hyperlink{struct_cosmology_ad8456f0eaa57c17e323994a2d0b1f159}{NLines}}
 \item 
-long \mbox{\hyperlink{struct_cosmology_a0e1e29102584fa127dceed1413913a13}{mode\+\_\+nu}}
+long \mbox{\hyperlink{struct_cosmology_a4b019ecf79b36a4e0ddc5e73f130f79a}{mode\+\_\+model}}
 \item 
-double \mbox{\hyperlink{struct_cosmology_ad00c7b692054c2201a342aca51eedabc}{cosmo\+\_\+pars}} \mbox{[}6\mbox{]}
+double \mbox{\hyperlink{struct_cosmology_afcdeb6e2220dcdfbfc948f81651a773f}{cosmo\+\_\+pars}} \mbox{[}\mbox{\hyperlink{header_8h_a4a80e098a43386a893ebc656b14e636c}{NPARS}}\mbox{]}
 \end{DoxyCompactItemize}
 
 
@@ -45,11 +48,11 @@ This includes, the \mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_
 \doxyparagraph{\texorpdfstring{ccs}{ccs}}
 {\footnotesize\ttfamily struct \mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_\+\+Cosmology\+\_\+\+Struct}} ccs}
 
-\mbox{\Hypertarget{struct_cosmology_ad00c7b692054c2201a342aca51eedabc}\label{struct_cosmology_ad00c7b692054c2201a342aca51eedabc}} 
+\mbox{\Hypertarget{struct_cosmology_afcdeb6e2220dcdfbfc948f81651a773f}\label{struct_cosmology_afcdeb6e2220dcdfbfc948f81651a773f}} 
 \index{Cosmology@{Cosmology}!cosmo\_pars@{cosmo\_pars}}
 \index{cosmo\_pars@{cosmo\_pars}!Cosmology@{Cosmology}}
 \doxyparagraph{\texorpdfstring{cosmo\_pars}{cosmo\_pars}}
-{\footnotesize\ttfamily double cosmo\+\_\+pars\mbox{[}6\mbox{]}}
[…]
