+++ docs/latex/d8/d43/ps__line__pt_8c.tex
@@ -6,7 +6,8 @@ Documented computation of Poisson shot noise and tree-\/level line power spectru
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:\nopagebreak
+Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -16,14 +17,16 @@ Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:\nopagebreak
 \doxysubsubsection*{Functions}
 \begin{DoxyCompactItemize}
 \item 
-double \mbox{\hyperlink{ps__line__pt_8c_adb0db1c5916a39f0b796a30b48b03f42}{PS\+\_\+line}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, size\+\_\+t line\+\_\+id)
+double \mbox{\hyperlink{ps__line__pt_8c_a5977cca9f8e0c388f2d9e6db7945b3dc}{PS\+\_\+tot\+\_\+interloper}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double mu, double z\+\_\+s, double z\+\_\+inter, size\+\_\+t line\+\_\+id)
+\begin{DoxyCompactList}\small\item\em Compute the 3D power spectrum of interloper line power spectrum in unit of micro K$^\wedge$2 Mpc$^\wedge$3. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{ps__line__pt_8c_a3c16139f4646a5084475aa3515d336a6}{PS\+\_\+line\+\_\+real}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, size\+\_\+t line\+\_\+id)
 \begin{DoxyCompactList}\small\item\em Compute the real-\/space 3D power spectrum of emission lines in unit of micro K$^\wedge$2 Mpc$^\wedge$3. \end{DoxyCompactList}\item 
 double \mbox{\hyperlink{ps__line__pt_8c_a2a5a8e5016f2d8728844564d93000751}{PS\+\_\+line\+\_\+\+RSD}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx\+\_\+ref, double k, double mu, double z, size\+\_\+t line\+\_\+id)
 \begin{DoxyCompactList}\small\item\em Compute the redshift-\/space 3D power spectrum of emission lines in unit of micro K$^\wedge$2 Mpc$^\wedge$3 as a function of wavenumber and angle w.\+r.\+t. \end{DoxyCompactList}\item 
 int \mbox{\hyperlink{ps__line__pt_8c_ae4ce1e29b505dae588eb2b8c8ae632e2}{ps\+\_\+line\+\_\+multipoles\+\_\+integrand}} (unsigned ndim, const double $\ast$x, void $\ast$p, unsigned fdim, double $\ast$fvalue)
-\begin{DoxyCompactList}\small\item\em Compute the multipole moments of redshift-\/space power spectrum of emission lines in unit of micro K$^\wedge$2 Mpc$^\wedge$3, integrating over the angle w.\+r.\+t LOS, weighted by. \end{DoxyCompactList}\item 
[…]
