+++ docs/latex/d0/d4a/structintegrand__parameters2.tex
@@ -11,11 +11,12 @@ Another structure passed to the integrators to hold the parameters fixed in the 
 
 
 
-Collaboration diagram for integrand\+\_\+parameters2\+:\nopagebreak
+Collaboration diagram for integrand\+\_\+parameters2\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=262pt]{df/d5f/structintegrand__parameters2__coll__graph}
+\includegraphics[width=336pt]{df/d5f/structintegrand__parameters2__coll__graph}
 \end{center}
 \end{figure}
 \doxysubsubsection*{Data Fields}
