+++ docs/latex/d0/d29/main_8c.tex
@@ -6,7 +6,8 @@ Documented main module, including functions to initilize and cleanup the cosmolo
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for main.\+c\+:\nopagebreak
+Include dependency graph for main.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -50,14 +51,16 @@ Free the memory allocated to cosmology structure and SFR interpolator.
 \begin{DoxyReturn}{Returns}
 void 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=350pt]{d0/d29/main_8c_a798f0df7ac8f26aae66b7364d7f4e6f0_cgraph}
+\includegraphics[width=268pt]{d0/d29/main_8c_a798f0df7ac8f26aae66b7364d7f4e6f0_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -90,8 +93,7 @@ In units of km/s
 
 omega\+\_\+b = Omega\+\_\+b h$^\wedge$2;
 
-3.\+0665Here is the call graph for this function\+:
-\nopagebreak
+3.\+0665Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
