+++ docs/latex/d7/d78/wnw__split_8c.tex
@@ -70,7 +70,8 @@ Compute the Eisentein-\/\+Hu approximate nowiggle component of linear matter pow
 \begin{DoxyReturn}{Returns}
 P\+\_\+nw(k) in unit of (Mpc)$^\wedge$3 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -109,7 +110,8 @@ Compute the Eisentein-\/\+Hu approximate wiggle component of linear matter power
 \begin{DoxyReturn}{Returns}
 P\+\_\+w(k) in unit of (Mpc)$^\wedge$3 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -220,7 +222,8 @@ approximate sound speed given in Eq. (26) of EH
 Eq. 21 of EH ref.
 
 Eq. 17 of EH ref ~\newline
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
