+++ docs/latex/dd/d24/survey__specs_8c.tex
@@ -67,11 +67,12 @@ We follow Giannantonio et al. to for determining kmax, and use gsl Brent solver 
 \begin{DoxyReturn}{Returns}
 kmax 
 \end{DoxyReturn}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=288pt]{dd/d24/survey__specs_8c_a1eb59c068328f2b2a2cad533600c0d78_icgraph}
+\includegraphics[width=286pt]{dd/d24/survey__specs_8c_a1eb59c068328f2b2a2cad533600c0d78_icgraph}
 \end{center}
 \end{figure}
 \mbox{\Hypertarget{survey__specs_8c_ab0f5b6af6991512fd864af10e6c2d54d}\label{survey__specs_8c_ab0f5b6af6991512fd864af10e6c2d54d}} 
@@ -143,11 +144,12 @@ Compute the comoving volume of a survey covering redshift up to z.
 \begin{DoxyReturn}{Returns}
 the comoving z-\/shell volume 
 \end{DoxyReturn}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=247pt]{dd/d24/survey__specs_8c_a91c4e0c62bcd3e179536fea62e8d53b5_icgraph}
+\includegraphics[width=246pt]{dd/d24/survey__specs_8c_a91c4e0c62bcd3e179536fea62e8d53b5_icgraph}
 \end{center}
 \end{figure}
 
