+++ docs/latex/dc/d8e/_i_r__res_8c.tex
@@ -206,11 +206,12 @@ Here is the call graph for this function\+:
 \includegraphics[width=350pt]{dc/d8e/_i_r__res_8c_a081df8016b0ce1b7160a3b49547e0348_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=336pt]{dc/d8e/_i_r__res_8c_a081df8016b0ce1b7160a3b49547e0348_icgraph}
+\includegraphics[width=334pt]{dc/d8e/_i_r__res_8c_a081df8016b0ce1b7160a3b49547e0348_icgraph}
 \end{center}
 \end{figure}
 \mbox{\Hypertarget{_i_r__res_8c_a6636c173929d53798037db6d86a74561}\label{_i_r__res_8c_a6636c173929d53798037db6d86a74561}} 
