+++ docs/latex/d0/d29/main_8c.tex
@@ -93,14 +93,16 @@ In units of km/s
 
 omega\+\_\+b = Omega\+\_\+b h$^\wedge$2;
 
-3.\+0665Here is the call graph for this function\+:\nopagebreak
+3.\+0665Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d0/d29/main_8c_a25a40b6614565f755233080a384c35f1_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
