+++ docs/latex/d8/d2a/ps__halo__1loop_8c.tex
@@ -107,7 +107,8 @@ Here is the call graph for this function\+:
 \includegraphics[width=350pt]{d8/d2a/ps__halo__1loop_8c_a06812b0e40a5e228647283536b82fa5a_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -150,11 +151,12 @@ Here is the call graph for this function\+:
 \includegraphics[width=350pt]{d8/d2a/ps__halo__1loop_8c_a05a2d154a1db42273c26d91b1145b892_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=304pt]{d8/d2a/ps__halo__1loop_8c_a05a2d154a1db42273c26d91b1145b892_icgraph}
+\includegraphics[width=301pt]{d8/d2a/ps__halo__1loop_8c_a05a2d154a1db42273c26d91b1145b892_icgraph}
 \end{center}
 \end{figure}
 \mbox{\Hypertarget{ps__halo__1loop_8c_af95342f8b89705476a09b2473195a749}\label{ps__halo__1loop_8c_af95342f8b89705476a09b2473195a749}} 
