+++ docs/latex/d7/db6/utilities_8c.tex
@@ -142,11 +142,12 @@ initialize a 1d array, with values in the range of \mbox{[}xmin,xmax\mbox{]} and
 \begin{DoxyReturn}{Returns}
 a pointer to a double type 1d array, with values initialized 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=268pt]{d7/db6/utilities_8c_aacaa803e04084752f13eb25e236f46aa_cgraph}
+\includegraphics[width=265pt]{d7/db6/utilities_8c_aacaa803e04084752f13eb25e236f46aa_cgraph}
 \end{center}
 \end{figure}
 Here is the caller graph for this function\+:
@@ -179,11 +180,12 @@ initialize a 1d array, with values in the range of \mbox{[}xmin,xmax\mbox{]} and
 \begin{DoxyReturn}{Returns}
 a pointer to a double type 1d array, with values initialized 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=293pt]{d7/db6/utilities_8c_abe221b42f18db326ebf3161fdcaa7f34_cgraph}
+\includegraphics[width=288pt]{d7/db6/utilities_8c_abe221b42f18db326ebf3161fdcaa7f34_cgraph}
 \end{center}
 \end{figure}
 \mbox{\Hypertarget{utilities_8c_aed3fb59a75f0f19ee1e7b860c0d5a8f8}\label{utilities_8c_aed3fb59a75f0f19ee1e7b860c0d5a8f8}} 
@@ -208,11 +210,12 @@ initialize a 1d array, with values in the range of \mbox{[}xmin,xmax\mbox{]} and
 \begin{DoxyReturn}{Returns}
 a pointer to a double type 1d array, with values initialized 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
-\includegraphics[width=282pt]{d7/db6/utilities_8c_aed3fb59a75f0f19ee1e7b860c0d5a8f8_cgraph}
+\includegraphics[width=278pt]{d7/db6/utilities_8c_aed3fb59a75f0f19ee1e7b860c0d5a8f8_cgraph}
 \end{center}
 \end{figure}
 Here is the caller graph for this function\+:
