+++ Include/wnw_split.h
@@ -1,4 +1,7 @@
 
+#ifndef WNW_SPLIT_H
+#define WNW_SPLIT_H
+
 double pk_nw_integrand(double x, void *par); 
 double pk_Gfilter_nw( struct Cosmology *Cx, double k, double kf0); 
 double EH_PS_w( struct Cosmology *Cx, double k, double k0, double p0);
@@ -7,6 +10,4 @@ double T0( struct Cosmology *Cx, double k);
 double T( struct Cosmology *Cx, double k);
 double Tt0( struct Cosmology *Cx, double k, double x1, double x2);
 
-
-
-
+#endif
