+++ Source/cosmology.c
@@ -1,3 +1,4 @@
+
 /** @file cosmology.c  Documented cosmology module 
  * 
  * Azadeh Moradinezhad Dizgah, November 4th 2021
@@ -14,7 +15,11 @@
  * -# Cosmology_free()              frees the memory allocated to cosmology structure
  * -# CL_Cosmology_initilize()      initializes the class cosmology structure
  * -# CL_Cosmology_free()           frees the class cosmology structure
+ * -# PS_xtrapol_init()             initizalizes the interpolator for Pm(k,z)
+ * -# PS_xtrapol_free()             frees the interpolator for Pm(k,z)
  * -# PS()                          computes matter power spectrum calling class and extrapolating if k>kmax computed by class
+ * -# PS_class()                    calls class to calculate 
+ * -# Pk_dlnPk_HV()                 reads in the mesaured Pm on HV simulation and interpolates 
  * -# transfer()                    computes matter transfer function calling class function
  * -# growth_D()                    computes the scale-dep growth factor
  * -# growth_f()                    computes the scale-dep growth rate dlnD(k,a)/dlna
@@ -24,20 +29,27 @@
  * -# angular_distance()            computes angular diamtere distance  using directly CLASS functions
  * -# comoving_radial_distance()    computes radial distance  using directly CLASS functions
  * -# sig_sq()                      computes variance of smoothed matter fluctuations
- * -# der_sig_sq()                  computes derivative of the variance of smoothed matter fluctuations w.r.t. smoothing scale
+ * -# der_lnsig_sq()                computes natural log derivative of the variance of smoothed matter fluctuations w.r.t. smoothing scale
  * -# sigma0_sq()                   computes variance of unsmoothed matter fluctuations
  * -# rhoc()                        computes the critical density of the universe 
  * -# R_scale()                     computes the size of a spherical halo corresponding to a given mass at z=0
  * -# R_scale_wrong()               computes the size of a spherical halo corresponding to a given mass at a given redshift
+ * -# R_vir()                       computes the virial radius
+ * -# concentration_cdm()           computes the halo concentration
+ * -# nfw profile()                 computes the nfw profile
  * -# window_rth()                  computes top-hat filter in real space
  * -# window_g()                    computes Gaussian window
  * -# window_kth()                  computes top-hat filter in Fourier space
  * -# derR_window_rth()             computes derivative of top-hat filter in real space w.r.t. smoothing scale 
  * -# derR_logwindow_g()            computes derivative of top-hat filter in Fourier space w.r.t. smoothing scale
+ *
  */
+
+
 
 #include "header.h"
[…]
