+++ docs/latex/d8/d43/ps__line__pt_8c.tex
@@ -6,7 +6,8 @@ Documented computation of Poisson shot noise and tree-\/level line power spectru
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:\nopagebreak
+Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -81,7 +82,8 @@ Compute the multipole moments of redshift-\/space power spectrum of emission lin
 P\+\_\+ell(k) ~\newline
  
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -111,14 +113,16 @@ The integrand function passed to hcubature integrator to compute the line power 
 {\em fvalue} & Input\+: Array of values of the integrand of dimension fdim return the error status \\
 \hline
 \end{DoxyParams}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d8/d43/ps__line__pt_8c_ae4ce1e29b505dae588eb2b8c8ae632e2_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -150,7 +154,8 @@ Compute the real-\/space 3D power spectrum of emission lines in unit of micro K$
 tree-\/level P\+\_\+clust(k) ~\newline
  
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -188,14 +193,16 @@ LOS
 tree-\/level P\+\_\+clust(k,mu) ~\newline
  
 \end{DoxyReturn}
-to plot the power spectrum in units of micro K$^\wedge$2 Mpc$^\wedge$3Here is the call graph for this function\+:\nopagebreak
+to plot the power spectrum in units of micro K$^\wedge$2 Mpc$^\wedge$3Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d8/d43/ps__line__pt_8c_a2a5a8e5016f2d8728844564d93000751_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -227,14 +234,16 @@ P\+\_\+poisson ~\newline
 \end{DoxyReturn}
 Boltzmann constant in unit of erg K$^\wedge$-\/1
 
-in unit of erg/sHere is the call graph for this function\+:\nopagebreak
+in unit of erg/sHere is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d8/d43/ps__line__pt_8c_aef342e8a0e97c247ed892896732d0dcd_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
[…]
