+++ docs/latex/dc/d8e/_i_r__res_8c.tex
@@ -6,8 +6,7 @@ Documented IR\+\_\+res module.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for IR\+\_\+res.\+c\+:
-\nopagebreak
+Include dependency graph for IR\+\_\+res.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -79,16 +78,14 @@ Compute the suppression factor IR\+\_\+sigma2.
 value of IR resummation suppression factor ~\newline
  
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{dc/d8e/_i_r__res_8c_ad57601a52710911744c3ab9ee4192075_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -116,16 +113,14 @@ Integrand to compute the suppression factor IR\+\_\+sigma2.
 integrand to be used in IR\+\_\+sigma2() function ~\newline
  
 \end{DoxyReturn}
-BAO\+\_\+scale = 110. Mpc/h.Here is the call graph for this function\+:
-\nopagebreak
+BAO\+\_\+scale = 110. Mpc/h.Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{dc/d8e/_i_r__res_8c_a96ac18bc55ca39bce830a35aac4f1dee_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -157,16 +152,14 @@ Compute the leading-\/order IR-\/resummed matter power spectrum, ala Ivanovic et
 value of leading IR-\/ressumed power spectrum ~\newline
  
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{dc/d8e/_i_r__res_8c_ac9b48827f9977633775bbff6971ec467_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -198,16 +191,14 @@ Compute the next-\/to-\/leading-\/order IR-\/resummed matter power spectrum, ala
 value of NL IR-\/ressumed power spectrum ~\newline
  
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{dc/d8e/_i_r__res_8c_a081df8016b0ce1b7160a3b49547e0348_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
[…]
