### 2022-01-20 bb426b0
removed doc folders
