+++ docs/latex/d8/d2a/ps__halo__1loop_8c.tex
@@ -1,377 +1,0 @@
-\hypertarget{ps__halo__1loop_8c}{}\doxysubsection{ps\+\_\+halo\+\_\+1loop.\+c File Reference}
-\label{ps__halo__1loop_8c}\index{ps\_halo\_1loop.c@{ps\_halo\_1loop.c}}
-
-
-Documented real-\/space, direct integration computation of 1loop contributions of the halo/galaxy power spectrum See ar\+Xiv\+:2010.\+14523 for explicit expressions.  
-
-
-{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+halo\+\_\+1loop.\+c\+:\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[width=350pt]{db/d05/ps__halo__1loop_8c__incl}
-\end{center}
-\end{figure}
-\doxysubsubsection*{Functions}
-\begin{DoxyCompactItemize}
-\item 
-double \mbox{\hyperlink{ps__halo__1loop_8c_a689859f9764b839323002fbb8be176e4}{PS\+\_\+hh\+\_\+G}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, double M, long mode\+\_\+pt, long IR\+\_\+switch, long SPLIT, long mode\+\_\+mf)
-\begin{DoxyCompactList}\small\item\em Compute the contributions up to 1loop to halo power spectrum for Gaussian initial conditions. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{ps__halo__1loop_8c_ad87e930ddfa85973919af4ca6b3d99ce}{PS\+\_\+hh\+\_\+\+PNG}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, double M, long mode\+\_\+pt, long IR\+\_\+switch, long SPLIT, long mode\+\_\+mf)
-\begin{DoxyCompactList}\small\item\em Compute contributions up to 1loop to halo power spectrum arising from non-\/\+Gaussian initial conditions of local shape. \end{DoxyCompactList}\item 
-static int \mbox{\hyperlink{ps__halo__1loop_8c_a9d475dd2ab5b0120a8c78ba15080494b}{G\+\_\+loop\+\_\+integrands}} (const int $\ast$ndim, const cubareal x\mbox{[}$\,$\mbox{]}, const int $\ast$ncomp, cubareal ff\mbox{[}$\,$\mbox{]}, void $\ast$p)
-\begin{DoxyCompactList}\small\item\em The integrand function passed passed to Cuhre integration routine to compute Gaussian 1loop terms. \end{DoxyCompactList}\item 
-void \mbox{\hyperlink{ps__halo__1loop_8c_a06812b0e40a5e228647283536b82fa5a}{Compute\+\_\+\+G\+\_\+loops}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, long IR\+\_\+switch, long hm\+\_\+switch, long SPLIT, double $\ast$result)
-\begin{DoxyCompactList}\small\item\em Compute the loop contributions dure to nonlinear evolution of matter fluctuations and nonlinear halo bias, present for Gaussian initial conditions. \end{DoxyCompactList}\item 
[…]
