+++ docs/latex/d6/da5/_global___structs_8h.tex
@@ -1,22 +1,0 @@
-\hypertarget{_global___structs_8h}{}\doxysubsection{Global\+\_\+\+Structs.\+h File Reference}
-\label{_global___structs_8h}\index{Global\_Structs.h@{Global\_Structs.h}}
-This graph shows which files directly or indirectly include this file\+:\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[width=350pt]{de/d32/_global___structs_8h__dep__incl}
-\end{center}
-\end{figure}
-\doxysubsubsection*{Data Structures}
-\begin{DoxyCompactItemize}
-\item 
-struct \mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_\+\+Cosmology\+\_\+\+Struct}}
-\begin{DoxyCompactList}\small\item\em Structure to store cosmology structure from CLASS-\/v3.\+1. \end{DoxyCompactList}\item 
-struct \mbox{\hyperlink{struct_cosmology}{Cosmology}}
-\begin{DoxyCompactList}\small\item\em Structure that holds varioud quantities that need to be evaluated for a given choice of cosmological paramteres. \end{DoxyCompactList}\item 
-struct \mbox{\hyperlink{struct_line}{Line}}
-\begin{DoxyCompactList}\small\item\em Structure that holds the Line-\/related quantities, including the interpolators for first and second moments of the line luminosity and the linear and quadratic luminosity-\/weighted line biases. \end{DoxyCompactList}\item 
-struct \mbox{\hyperlink{struct_p_s__xtr}{PS\+\_\+xtr}}
-\item 
-struct \mbox{\hyperlink{structglobals}{globals}}
-\begin{DoxyCompactList}\small\item\em A global structure including the values of cosmological parmaeters, 2d interpolator of SFR, and names of various files. \end{DoxyCompactList}\end{DoxyCompactItemize}
