+++ docs/latex/dc/d8e/_i_r__res_8c.tex
@@ -1,367 +1,0 @@
-\hypertarget{_i_r__res_8c}{}\doxysubsection{IR\+\_\+res.\+c File Reference}
-\label{_i_r__res_8c}\index{IR\_res.c@{IR\_res.c}}
-
-
-Documented IR\+\_\+res module.  
-
-
-{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for IR\+\_\+res.\+c\+:\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[width=350pt]{d2/d5b/_i_r__res_8c__incl}
-\end{center}
-\end{figure}
-\doxysubsubsection*{Functions}
-\begin{DoxyCompactItemize}
-\item 
-double \mbox{\hyperlink{_i_r__res_8c_ac9b48827f9977633775bbff6971ec467}{pm\+\_\+\+IR\+\_\+\+LO}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, long SPLIT)
-\begin{DoxyCompactList}\small\item\em Compute the leading-\/order IR-\/resummed matter power spectrum, ala Ivanovic et al. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{_i_r__res_8c_a081df8016b0ce1b7160a3b49547e0348}{pm\+\_\+\+IR\+\_\+\+NLO}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, long SPLIT)
-\begin{DoxyCompactList}\small\item\em Compute the next-\/to-\/leading-\/order IR-\/resummed matter power spectrum, ala Ivanovic et al. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{_i_r__res_8c_a96ac18bc55ca39bce830a35aac4f1dee}{IR\+\_\+\+Sigma2\+\_\+integrand}} (double x, void $\ast$par)
-\begin{DoxyCompactList}\small\item\em Integrand to compute the suppression factor IR\+\_\+sigma2. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{_i_r__res_8c_ad57601a52710911744c3ab9ee4192075}{IR\+\_\+\+Sigma2}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double z, double k0, long SPLIT)
-\begin{DoxyCompactList}\small\item\em Compute the suppression factor IR\+\_\+sigma2. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{_i_r__res_8c_a6636c173929d53798037db6d86a74561}{pm\+\_\+nowiggle}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, double kf0, int \mbox{\hyperlink{header_8h_a4b66d5e31b5dc18b314c8a68163263bd}{cleanup}}, long SPLIT)
-\begin{DoxyCompactList}\small\item\em Compute the no-\/wiggle componenet of the matter power spectrum. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{_i_r__res_8c_acaaad0d343ab0665ec92102ca3b9fa8f}{pm\+\_\+nowiggle\+\_\+bspline}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, int \mbox{\hyperlink{header_8h_a4b66d5e31b5dc18b314c8a68163263bd}{cleanup}})
[…]
