+++ docs/latex/d7/d78/wnw__split_8c.tex
@@ -1,0 +1,292 @@
+\hypertarget{wnw__split_8c}{}\doxysubsection{wnw\+\_\+split.\+c File Reference}
+\label{wnw__split_8c}\index{wnw\_split.c@{wnw\_split.c}}
+
+
+Documented wiggle-\/nowiggle split based on 3d Gaussian filter in linear k, and using the Eisentein-\/\+Hu wiggle-\/no wiggle template ar\+Xiv\+:astro-\/ph/9709112.  
+
+
+{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
+Include dependency graph for wnw\+\_\+split.\+c\+:\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=350pt]{d0/d12/wnw__split_8c__incl}
+\end{center}
+\end{figure}
+\doxysubsubsection*{Functions}
+\begin{DoxyCompactItemize}
+\item 
+double \mbox{\hyperlink{wnw__split_8c_a1d4fa9e8e5258809f84b0e2150641758}{pk\+\_\+\+Gfilter\+\_\+nw}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double k0)
+\begin{DoxyCompactList}\small\item\em Compute the nowiggle component of linear matter power spectrum using 3d Gaussian filter Computing the nowiggle component involves calculating an integral (Eq. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{wnw__split_8c_ad6bd78b923d93da921f7f86f956d2d77}{pk\+\_\+nw\+\_\+integrand}} (double x, void $\ast$par)
+\begin{DoxyCompactList}\small\item\em Integrand to compute the nowiggle matter power spectrum. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{wnw__split_8c_a0cb6b540dc962bfd28da54fa6819f4df}{EH\+\_\+\+PS\+\_\+w}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double k0, double pk0)
+\begin{DoxyCompactList}\small\item\em Compute the Eisentein-\/\+Hu approximate wiggle component of linear matter power spectrum. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{wnw__split_8c_a0859675908f887c5e50061678261bf4b}{EH\+\_\+\+PS\+\_\+nw}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double k0, double pk0)
+\begin{DoxyCompactList}\small\item\em Compute the Eisentein-\/\+Hu approximate nowiggle component of linear matter power spectrum. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{wnw__split_8c_ab920120944723a06efde25145e11a333}{T0}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k)
+\begin{DoxyCompactList}\small\item\em Compute the no-\/baryon transfer function given in Eq. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{wnw__split_8c_ae73af41bcbfa0911ec17d4ac2d91baa3}{T}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k)
[…]
