+++ docs/latex/refman.tex
@@ -174,8 +174,6 @@
   % re-enable anchors again
   \hypersetup{pageanchor=true}
 %--- Begin generated contents ---
-\doxysection{Summary}
-\label{index}\hypertarget{index}{}\input{index}
 \doxysection{Data Structure Index}
 \input{annotated}
 \doxysection{File Index}
@@ -189,7 +187,6 @@
 \input{d2/d9d/struct_line}
 \input{d0/d0c/struct_p_s__xtr}
 \doxysection{File Documentation}
-\input{d9/dd6/_r_e_a_d_m_e_8md}
 \input{d6/da5/_global___structs_8h}
 \input{d6/da5/_global___structs_8h_source}
 \input{df/dcb/header_8h}
