### 2022-01-20 b05b2d2
Update github-actions-demo.yml
