+++ .github/workflows/github-actions-demo.yml
@@ -13,7 +13,7 @@ jobs:
     - name: Doxygen Action
       uses: mattnotmitt/doxygen-action@v1.1.0
       with:
-          doxyfile-path: ./Doxyfile
+          doxyfile-path: ./docs/Doxyfile
           working-directory: .
     
     # Deploy the HTML documentation to GitHub Pages
@@ -21,7 +21,7 @@ jobs:
       uses: peaceiris/actions-gh-pages@v3
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: ./doc/html/
+        publish_dir: ./docs/html/
         enable_jekyll: false
         allow_empty_commit: false
         force_orphan: true
