+++ .github/workflows/github-actions-demo.yml
@@ -2,31 +2,7 @@
 
 on: [push]
 
-jobs:
-  build-documentation:
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    
-    # Build the HTML documentation
-    - name: Doxygen Action
-      uses: mattnotmitt/doxygen-action@v1.1.0
-      with:
-          doxyfile-path: ./docs/Doxyfile
-          working-directory: .
-    
-    # Deploy the HTML documentation to GitHub Pages
-    - name: GH Pages Deployment
-      uses: peaceiris/actions-gh-pages@v3
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: ./docs/html/
-        enable_jekyll: false
-        allow_empty_commit: false
-        force_orphan: true
-        publish_branch: gh-pages
-        
+jobs:     
   paper:
     runs-on: ubuntu-latest
     name: Paper Draft
