+++ docs/latex/files.tex
@@ -1,15 +1,0 @@
-\doxysubsection{File List}
-Here is a list of all files with brief descriptions\+:\begin{DoxyCompactList}
-\item\contentsline{section}{\mbox{\hyperlink{_global___structs_8h}{Global\+\_\+\+Structs.\+h}} }{\pageref{_global___structs_8h}}{}
-\item\contentsline{section}{\mbox{\hyperlink{header_8h}{header.\+h}} }{\pageref{header_8h}}{}
-\item\contentsline{section}{\mbox{\hyperlink{cosmology_8c}{cosmology.\+c}} \\*Documented cosmology module }{\pageref{cosmology_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{_i_r__res_8c}{IR\+\_\+res.\+c}} \\*Documented IR\+\_\+res module }{\pageref{_i_r__res_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{line__ingredients_8c}{line\+\_\+ingredients.\+c}} \\*Documented line\+\_\+ingredients module }{\pageref{line__ingredients_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{main_8c}{main.\+c}} \\*Documented main module, including functions to initilize and cleanup the cosmology structure and examples of calls to functions in other modules to compute the line clustering and shot power spectrum }{\pageref{main_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{ps__halo__1loop_8c}{ps\+\_\+halo\+\_\+1loop.\+c}} \\*Documented real-\/space, direct integration computation of 1loop contributions of the halo/galaxy power spectrum See ar\+Xiv\+:2010.\+14523 for explicit expressions }{\pageref{ps__halo__1loop_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{ps__line__hm_8c}{ps\+\_\+line\+\_\+hm.\+c}} \\*Documented halo-\/model computation of line power spectrum, including clustering and stochastic contributions beyond Poisson limit }{\pageref{ps__line__hm_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{ps__line__pt_8c}{ps\+\_\+line\+\_\+pt.\+c}} \\*Documented computation of Poisson shot noise and tree-\/level line power spectrum in real and redshift-\/space }{\pageref{ps__line__pt_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{survey__specs_8c}{survey\+\_\+specs.\+c}} \\*Documented computation of some survey-\/related functions }{\pageref{survey__specs_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{utilities_8c}{utilities.\+c}} \\*Documented basic utility functions used by other modules of the code }{\pageref{utilities_8c}}{}
-\item\contentsline{section}{\mbox{\hyperlink{wnw__split_8c}{wnw\+\_\+split.\+c}} \\*Documented wiggle-\/nowiggle split based on 3d Gaussian filter in linear k, and using the Eisentein-\/\+Hu wiggle-\/no wiggle template ar\+Xiv\+:astro-\/ph/9709112 }{\pageref{wnw__split_8c}}{}
-\end{DoxyCompactList}

[…]
