+++ docs/latex/d7/d10/line__ingredients_8c.tex
@@ -1,0 +1,1041 @@
+\hypertarget{line__ingredients_8c}{}\doxysubsection{line\+\_\+ingredients.\+c File Reference}
+\label{line__ingredients_8c}\index{line\_ingredients.c@{line\_ingredients.c}}
+
+
+Documented line\+\_\+ingredients module.  
+
+
+{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
+Include dependency graph for line\+\_\+ingredients.\+c\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=350pt]{dc/d32/line__ingredients_8c__incl}
+\end{center}
+\end{figure}
+\doxysubsubsection*{Functions}
+\begin{DoxyCompactItemize}
+\item 
+struct \mbox{\hyperlink{struct_line}{Line}} $\ast$ \mbox{\hyperlink{line__ingredients_8c_aa4696edc9829e9a3c2c1012a3e9d9dbe}{Line\+\_\+alloc\+\_\+init}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, long line\+\_\+type, size\+\_\+t npoints\+\_\+interp, double M\+\_\+min, long mode\+\_\+mf)
+\begin{DoxyCompactList}\small\item\em Allocate the memory and initialize the the line structure. \end{DoxyCompactList}\item 
+int \mbox{\hyperlink{line__ingredients_8c_a255599f7e846dd569f4ce4c946157dec}{Line\+\_\+free}} (struct \mbox{\hyperlink{struct_line}{Line}} $\ast$Lx)
+\begin{DoxyCompactList}\small\item\em Free the line structure. \end{DoxyCompactList}\item 
+int \mbox{\hyperlink{line__ingredients_8c_a0b51100183117f08ce0cfa9428fa72e4}{Line\+\_\+evaluate}} (struct \mbox{\hyperlink{struct_line}{Line}} $\ast$Lx, double $\ast$zz, double $\ast$res)
+\begin{DoxyCompactList}\small\item\em Allocate the memory and initialize the the line structure. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{line__ingredients_8c_ae44a16b587a91c0ea92dea263d20a5be}{mult\+\_\+func}} (double sigma, long mode\+\_\+mf)
+\begin{DoxyCompactList}\small\item\em Compute the multiplicity function needed to compute the halo mass function Three models are implemented\+: Press-\/\+Schechter, Sheth-\/\+Tormen and Tinker see Pillepich et al arxiv\+: 0811.\+4176 for the expressions. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{line__ingredients_8c_a2ca7a5add6fca8bd94165085f351f702}{mass\+\_\+func}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double M, double z, long mode\+\_\+mf)
+\begin{DoxyCompactList}\small\item\em Compute the halo mass function for Press-\/\+Schechter, Sheth-\/\+Tormen and Tinker models see Pillepich et al arxiv\+: 0811.\+4176 for the expressions. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{line__ingredients_8c_a60bf4840dc0204947ec426719979065e}{mass\+\_\+func\+\_\+sims}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double M, double z, long mode\+\_\+mf)
[…]
