+++ docs/latex/d2/d9d/struct_line.tex
@@ -1,116 +1,0 @@
-\hypertarget{struct_line}{}\doxysubsection{Line Struct Reference}
-\label{struct_line}\index{Line@{Line}}
-
-
-Structure that holds the Line-\/related quantities, including the interpolators for first and second moments of the line luminosity and the linear and quadratic luminosity-\/weighted line biases.  
-
-
-
-
-{\ttfamily \#include $<$Global\+\_\+\+Structs.\+h$>$}
-
-\doxysubsubsection*{Data Fields}
-\begin{DoxyCompactItemize}
-\item 
-long \mbox{\hyperlink{struct_line_a78c3e4b98961dce5a733adbc398238c9}{Line\+Type}}
-\item 
-int \mbox{\hyperlink{struct_line_ad06983e7f6e71b233ea7ff3dee1952f2}{initialized}}
-\item 
-size\+\_\+t \mbox{\hyperlink{struct_line_a5af76072d7991fda41ed345aa75ee7ee}{npoints\+Interp}}
-\item 
-double \mbox{\hyperlink{struct_line_a2537e68f750a3936366084bc7d564575}{line\+\_\+freq}}
-\item 
-gsl\+\_\+interp\+\_\+accel $\ast$ \mbox{\hyperlink{struct_line_ad2e644b5a90d727f8e08a9eb0d708f80}{mom1\+\_\+accel\+\_\+ptr}}
-\item 
-gsl\+\_\+spline $\ast$ \mbox{\hyperlink{struct_line_ae8040cc245eea740519f6461096385df}{mom1\+\_\+spline\+\_\+ptr}}
-\item 
-gsl\+\_\+interp\+\_\+accel $\ast$ \mbox{\hyperlink{struct_line_ac33bc84e6c67b56e4076354c2a35fb10}{mom2\+\_\+accel\+\_\+ptr}}
-\item 
-gsl\+\_\+spline $\ast$ \mbox{\hyperlink{struct_line_aa116bd0834a03b6b4f35ca1e999adcbc}{mom2\+\_\+spline\+\_\+ptr}}
-\item 
-gsl\+\_\+interp\+\_\+accel $\ast$ \mbox{\hyperlink{struct_line_aa2f7d89582262545df111c44c1af1e5e}{b1\+\_\+\+LW\+\_\+accel\+\_\+ptr}}
-\item 
-gsl\+\_\+spline $\ast$ \mbox{\hyperlink{struct_line_a44f30751fd31a8e0aad7c43cb4888195}{b1\+\_\+\+LW\+\_\+spline\+\_\+ptr}}
-\item 
-gsl\+\_\+interp\+\_\+accel $\ast$ \mbox{\hyperlink{struct_line_ac29bb17922e47d34f72c4f1c5b6098f4}{b2\+\_\+\+LW\+\_\+accel\+\_\+ptr}}
-\item 
-gsl\+\_\+spline $\ast$ \mbox{\hyperlink{struct_line_ab5b6d7ccfe895ebf565dab2c766ed15e}{b2\+\_\+\+LW\+\_\+spline\+\_\+ptr}}
-\end{DoxyCompactItemize}
-
-
-\doxysubsubsection{Detailed Description}
-Structure that holds the Line-\/related quantities, including the interpolators for first and second moments of the line luminosity and the linear and quadratic luminosity-\/weighted line biases. 
-
-\doxysubsubsection{Field Documentation}
[…]
