+++ docs/latex/d6/d32/cosmology_8c.tex
@@ -1,1194 +1,0 @@
-\hypertarget{cosmology_8c}{}\doxysubsection{cosmology.\+c File Reference}
-\label{cosmology_8c}\index{cosmology.c@{cosmology.c}}
-
-
-Documented cosmology module.  
-
-
-{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for cosmology.\+c\+:
-\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[width=350pt]{d9/da3/cosmology_8c__incl}
-\end{center}
-\end{figure}
-\doxysubsubsection*{Functions}
-\begin{DoxyCompactItemize}
-\item 
-int \mbox{\hyperlink{cosmology_8c_a4b4377cdc8e3bc9374e420bf65b5ce90}{Cosmology\+\_\+init}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double pk\+\_\+kmax, double pk\+\_\+zmax, int nlines, int $\ast$line\+\_\+type, size\+\_\+t npoints\+\_\+interp, double M\+\_\+min, long mode\+\_\+mf)
-\begin{DoxyCompactList}\small\item\em Allocate memory and initialize the cosmology structure, which includes the CLASS cosmology structure and line strucrure. \end{DoxyCompactList}\item 
-int \mbox{\hyperlink{cosmology_8c_a8ddfed4a843beb762df124badc47011e}{Cosmology\+\_\+free}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx)
-\begin{DoxyCompactList}\small\item\em Free the memory allocated to cosmology structure. \end{DoxyCompactList}\item 
-int \mbox{\hyperlink{cosmology_8c_a6cb2155e541823871af839a39ccec1e7}{CL\+\_\+\+Cosmology\+\_\+initilize}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double pk\+\_\+kmax, double pk\+\_\+zmax)
-\begin{DoxyCompactList}\small\item\em Allocate memory and initialize the CLASS cosmology structure. \end{DoxyCompactList}\item 
-int \mbox{\hyperlink{cosmology_8c_a9648c21ce0e6371a11de7a0f0ffec493}{CL\+\_\+\+Cosmology\+\_\+free}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx)
-\begin{DoxyCompactList}\small\item\em Free the memory allocated to CLASS cosmology structure. \end{DoxyCompactList}\item 
-struct \mbox{\hyperlink{struct_p_s__xtr}{PS\+\_\+xtr}} $\ast$ \mbox{\hyperlink{cosmology_8c_a9e22a5bbc8d0f80cafea3523b2432c82}{PS\+\_\+xtrapol\+\_\+init}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx)
-\begin{DoxyCompactList}\small\item\em Call CLASS and build a 2d interpolator for Plin(k,z), which will be used when building an extrapolator for P(k,z) beyond a given kmax computed by CLASS. \end{DoxyCompactList}\item 
-int \mbox{\hyperlink{cosmology_8c_a9f048ed16d14d0f8d8970524b2e6e57c}{PS\+\_\+xtrapol\+\_\+free}} (struct \mbox{\hyperlink{struct_p_s__xtr}{PS\+\_\+xtr}} $\ast$PS\+\_\+xtrapol)
[…]

[…]
