+++ docs/latex/index.tex
@@ -1,0 +1,33 @@
+\label{index_md_README}%
+\Hypertarget{index_md_README}%
+ Author\+: Azadeh Moradinezhad Dizgah ~\newline
+
+
+with contributions from Farnik Nikakhtar
+
+Welcom to lim\+Halo\+PT, a numerical package for computing the clustering and shot-\/noise contributions to the power spectrum of line intensity/temprature fluctuations within halo-\/model framework. The current version of the code, is limited to real-\/space, and redshift-\/space distortions will be included in the next release.
+
+The extended halo model of line intensity power spectrum implemented in lim\+Halo\+PT, combines the predictions of EFTof\+LSS for halo power spectrum with the standard halo model to account for the nonlinear evolution of matter fluctuations and the nonlinear biasing relation between line intensity fluctuations and the underlying dark matter distribution in 2-\/halo term. Furthermore, the model includes the effect of large bulk velocities (Infrared Resummation) in the 2-\/halo term. The deviations from Poisson shot noise on large scales are also computed within the halo model.
+
+This package is released together with the following publication, \href{https://arxiv.org/abs/2111.03717}{\texttt{ ar\+Xiv\+:2111.\+03717}}, where the prediction of the model are tested against new suite of simulated intensity (brightness temprature) maps of CO and \mbox{[}CII\mbox{]} lines. The mesheded fileds from Mithra\+LIMSims will be publically avilable on \href{http://cyril.astro.berkeley.edu/MithraLIMSims}{\texttt{ Mithra\+LIMSims}}. The code to analyse the simulated maps is an extension of the toolskit used in analysin Hidden-\/\+Valley simulations, and is publically avialble on \href{https://github.com/farnikn/MithraLIMSims}{\texttt{ LIM Analysis}}. As discussed in the manuscript above, the packages to compute the theory predictions and creating the simulated intensity maps can be straightforwardly extended to compute the power spectrum signal of other emission lines (emitted from star-\/froming galaxies), beside CO and \mbox{[}CII\mbox{]}.
+
+The source code of this package is publically avaialbel on Git\+Hub at \href{https://github.com/amoradinejad/limHaloPT}{\texttt{ lim\+Halo\+PT}}. ~\newline
+ ~\hypertarget{index_autotoc_md5}{}\doxysubsubsection{Dependencies}\label{index_autotoc_md5}
+The lim\+Halo\+PT package calls various functions from \href{https://github.com/lesgourg/class_public}{\texttt{ CLASS}} Boltzman solver, including the matter power spectrum and transfer functions, growth factor etc. Therefore, you need to first download and compile CLASS code, and place the \char`\"{}libclass.\+a\char`\"{} file in the \char`\"{} CLASS/lib/\char`\"{} folder. Furtehrmore, the loop calculations are performed with direct numerical integration, using routines of \href{http://www.feynarts.de/cuba/}{\texttt{ CUBA}} library. Furthermore, the code heavily uses functions of \href{https://www.gnu.org/software/gsl/doc/html/}{\texttt{ GSL}} scientific library. Therfore, make sure that the two libraries are correctly linked to lim\+Halo\+PT by making necassary modifcations to the makefile (placed in Source directory) of lim\+Halo\+PT package. ~\newline
+ ~\hypertarget{index_autotoc_md6}{}\doxysubsubsection{Compilation and Usage}\label{index_autotoc_md6}
+
+\begin{DoxyItemize}
+\item To compile, type\+: make ~\newline
+
[…]
