+++ docs/latex/dir_856524284ebe840938865dc061f982fb.tex
@@ -1,0 +1,9 @@
+\hypertarget{dir_856524284ebe840938865dc061f982fb}{}\doxysubsection{Include Directory Reference}
+\label{dir_856524284ebe840938865dc061f982fb}\index{Include Directory Reference@{Include Directory Reference}}
+\doxysubsubsection*{Files}
+\begin{DoxyCompactItemize}
+\item 
+file \mbox{\hyperlink{_global___structs_8h}{Global\+\_\+\+Structs.\+h}}
+\item 
+file \mbox{\hyperlink{header_8h}{header.\+h}}
+\end{DoxyCompactItemize}

[…]
