+++ Include/header.h
@@ -110,13 +110,13 @@
 #include "../Class/include/class.h"
 /// \endcond
 #include "cubature.h"
-#include "Global_Structs.h"
+#include "global_structs.h"
 #include "utilities.h"
 #include "cosmology.h"
 #include "survey_specs.h"
 #include "line_ingredients.h"	
 #include "wnw_split.h"
-#include "IR_res.h"
+#include "ir_res.h"
 #include "ps_halo_1loop.h"
 #include "ps_line_pt.h"
 #include "ps_line_hm.h"
