### 2022-01-21 7fc31f8
Update and rename IR_res.c to ir_res.c
