+++ docs/latex/dd/d24/survey__specs_8c.tex
@@ -1,180 +1,0 @@
-\hypertarget{survey__specs_8c}{}\doxysubsection{survey\+\_\+specs.\+c File Reference}
-\label{survey__specs_8c}\index{survey\_specs.c@{survey\_specs.c}}
-
-
-Documented computation of some survey-\/related functions.  
-
-
-{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for survey\+\_\+specs.\+c\+:\nopagebreak
-\begin{figure}[H]
-\begin{center}
-\leavevmode
-\includegraphics[width=350pt]{d9/d56/survey__specs_8c__incl}
-\end{center}
-\end{figure}
-\doxysubsubsection*{Functions}
-\begin{DoxyCompactItemize}
-\item 
-double \mbox{\hyperlink{survey__specs_8c_a91c4e0c62bcd3e179536fea62e8d53b5}{shell\+\_\+volume}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double z, double fsky)
-\begin{DoxyCompactList}\small\item\em Compute the comoving volume of a survey covering redshift up to z. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{survey__specs_8c_a324eb852d8e2bb534361c033279c957c}{kmin\+\_\+val}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double zmin, double zmax, double fsky)
-\begin{DoxyCompactList}\small\item\em Compute the size of fundumental mode corresponding to the comoving volume enclosed in a redshift bin \mbox{[}zmin,zmax\mbox{]}. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{survey__specs_8c_a1eb59c068328f2b2a2cad533600c0d78}{kmax\+\_\+\+Brent}} (double kmax, void $\ast$params)
-\begin{DoxyCompactList}\small\item\em The function to be passed to Brent solver to compute kmax. \end{DoxyCompactList}\item 
-double \mbox{\hyperlink{survey__specs_8c_ab0f5b6af6991512fd864af10e6c2d54d}{kmax\+\_\+\+Brent\+\_\+solver}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double z)
-\begin{DoxyCompactList}\small\item\em Compute the maximum k-\/value used in Fisher forecast at each redshift bin. \end{DoxyCompactList}\end{DoxyCompactItemize}
-\doxysubsubsection*{Variables}
-\begin{DoxyCompactItemize}
-\item 
-struct \mbox{\hyperlink{structglobals}{globals}} \mbox{\hyperlink{survey__specs_8c_a61c6674f2b410e126a682325ce5360c0}{gb}}
-\end{DoxyCompactItemize}
-
-
-\doxysubsubsection{Detailed Description}
-Documented computation of some survey-\/related functions. 
-
-Azadeh Moradinezhad Dizgah, November 4th 2021
-
-In summary, the following functions can be called from other modules\+:
-\begin{DoxyEnumerate}
-\item \mbox{\hyperlink{survey__specs_8c_a91c4e0c62bcd3e179536fea62e8d53b5}{shell\+\_\+volume()}} computes the comoving volume of a survey covering redshift up to z
[…]
