+++ docs/latex/refman.tex
@@ -1,215 +1,0 @@
-% Latex header for doxygen 1.9.2
-  % Handle batch mode
-  % to overcome problems with too many open files
-  \let\mypdfximage\pdfximage\def\pdfximage{\immediate\mypdfximage}
-  % Set document class depending on configuration
-  \documentclass[twoside]{article}
-  %% moved from doxygen.sty due to workaround for LaTex 2019 version and unmaintained tabu package
-  \usepackage{ifthen}
-  \ifx\requestedLaTeXdate\undefined
-    \usepackage{array}
-  \else
-    \usepackage{array}[=2016-10-06]
-  \fi
-  %%
-  % Packages required by doxygen
-  \usepackage{fixltx2e} % for \textsubscript
-  \usepackage{doxygen}
-  \usepackage{doxygen}
-  \usepackage{graphicx}
-  \usepackage[utf8]{inputenc}
-  \usepackage{makeidx}
-  \PassOptionsToPackage{warn}{textcomp}
-  \usepackage{textcomp}
-  \usepackage[nointegrals]{wasysym}
-  \usepackage{ifxetex}
-  % NLS support packages
-  % Define default fonts
-  % Font selection
-  \usepackage[T1]{fontenc}
-  % set main and monospaced font
-  \usepackage[scaled=.90]{helvet}
-\usepackage{courier}
-\renewcommand{\familydefault}{\sfdefault}
-  \usepackage{sectsty}
-  \allsectionsfont{%
-    \fontseries{bc}\selectfont%
-    \color{darkgray}%
-  }
-  \renewcommand{\DoxyLabelFont}{%
-    \fontseries{bc}\selectfont%
-    \color{darkgray}%
-  }
-  \newcommand{\+}{\discretionary{\mbox{\scriptsize$\hookleftarrow$}}{}{}}
-   % Arguments of doxygenemoji:
-   % 1) ':<text>:' form of the emoji, already LaTeX-escaped
-   % 2) file with the name of the emoji without the .png extension
-   % in case image exist use this otherwise use the ':<text>:' form
-   \newcommand{\doxygenemoji}[2]{%
-     \IfFileExists{./#2.png}{\raisebox{-0.1em}{\includegraphics[height=0.9em]{./#2.png}}}{#1}%
-   }
-  % Page & text layout
-  \usepackage{geometry}
-  \geometry{%
-    a4paper,%
-    top=2.5cm,%
-    bottom=2.5cm,%
-    left=2.5cm,%
-    right=2.5cm%
-  }
-  % Allow a bit of overflow to go unnoticed by other means
-  \tolerance=750
-  \hfuzz=15pt
-  \hbadness=750
-  \setlength{\emergencystretch}{15pt}
-  \setlength{\parindent}{0cm}
-  \newcommand{\doxynormalparskip}{\setlength{\parskip}{3ex plus 2ex minus 2ex}}
-  \newcommand{\doxytocparskip}{\setlength{\parskip}{1ex plus 0ex minus 0ex}}
-  \doxynormalparskip
-  % Redefine paragraph/subparagraph environments, using sectsty fonts
-  \makeatletter
-  \renewcommand{\paragraph}{%
-    \@startsection{paragraph}{4}{0ex}{-1.0ex}{1.0ex}{%
-      \normalfont\normalsize\bfseries\SS@parafont%
-    }%
-  }
-  \renewcommand{\subparagraph}{%
-    \@startsection{subparagraph}{5}{0ex}{-1.0ex}{1.0ex}{%
-      \normalfont\normalsize\bfseries\SS@subparafont%
-    }%
-  }
-  \makeatother
-  \makeatletter
[…]
