+++ docs/latex/annotated.tex
@@ -1,10 +1,0 @@
-\doxysubsection{Data Structures}
-Here are the data structures with brief descriptions\+:\begin{DoxyCompactList}
-\item\contentsline{section}{\mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_\+\+Cosmology\+\_\+\+Struct}} \\*Structure to store cosmology structure from CLASS-\/v3.\+1 }{\pageref{struct_class___cosmology___struct}}{}
-\item\contentsline{section}{\mbox{\hyperlink{struct_cosmology}{Cosmology}} \\*Structure that holds varioud quantities that need to be evaluated for a given choice of cosmological paramteres }{\pageref{struct_cosmology}}{}
-\item\contentsline{section}{\mbox{\hyperlink{structglobals}{globals}} \\*A global structure including the values of cosmological parmaeters, 2d interpolator of SFR, and names of various files }{\pageref{structglobals}}{}
-\item\contentsline{section}{\mbox{\hyperlink{structintegrand__parameters}{integrand\+\_\+parameters}} \\*A structure passed to the integrators to hold the parameters fixed in the integration }{\pageref{structintegrand__parameters}}{}
-\item\contentsline{section}{\mbox{\hyperlink{structintegrand__parameters2}{integrand\+\_\+parameters2}} \\*Another structure passed to the integrators to hold the parameters fixed in the integration }{\pageref{structintegrand__parameters2}}{}
-\item\contentsline{section}{\mbox{\hyperlink{struct_line}{Line}} \\*Structure that holds the Line-\/related quantities, including the interpolators for first and second moments of the line luminosity and the linear and quadratic luminosity-\/weighted line biases }{\pageref{struct_line}}{}
-\item\contentsline{section}{\mbox{\hyperlink{struct_p_s__xtr}{PS\+\_\+xtr}} }{\pageref{struct_p_s__xtr}}{}
-\end{DoxyCompactList}

[…]
