+++ docs/latex/d2/dbd/struct_cosmology.tex
@@ -1,0 +1,80 @@
+\hypertarget{struct_cosmology}{}\doxysubsection{Cosmology Struct Reference}
+\label{struct_cosmology}\index{Cosmology@{Cosmology}}
+
+
+Structure that holds varioud quantities that need to be evaluated for a given choice of cosmological paramteres.  
+
+
+
+
+{\ttfamily \#include $<$Global\+\_\+\+Structs.\+h$>$}
+
+
+
+Collaboration diagram for Cosmology\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=336pt]{d3/df4/struct_cosmology__coll__graph}
+\end{center}
+\end{figure}
+\doxysubsubsection*{Data Fields}
+\begin{DoxyCompactItemize}
+\item 
+struct \mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_\+\+Cosmology\+\_\+\+Struct}} \mbox{\hyperlink{struct_cosmology_af8e7e270adf2573ca16f2037d580cb5b}{ccs}}
+\item 
+struct \mbox{\hyperlink{struct_line}{Line}} $\ast$$\ast$ \mbox{\hyperlink{struct_cosmology_a0f59f743170aa24dbaeebd4058ac226a}{Lines}}
+\item 
+struct \mbox{\hyperlink{struct_p_s__xtr}{PS\+\_\+xtr}} $\ast$ \mbox{\hyperlink{struct_cosmology_a5dee8f06ef55f61fcf2ad920f4f80e65}{PS\+\_\+xtrapol}}
+\item 
+int \mbox{\hyperlink{struct_cosmology_ad8456f0eaa57c17e323994a2d0b1f159}{NLines}}
+\item 
+long \mbox{\hyperlink{struct_cosmology_a4b019ecf79b36a4e0ddc5e73f130f79a}{mode\+\_\+model}}
+\item 
+double \mbox{\hyperlink{struct_cosmology_afcdeb6e2220dcdfbfc948f81651a773f}{cosmo\+\_\+pars}} \mbox{[}\mbox{\hyperlink{header_8h_a4a80e098a43386a893ebc656b14e636c}{NPARS}}\mbox{]}
+\end{DoxyCompactItemize}
+
+
+\doxysubsubsection{Detailed Description}
+Structure that holds varioud quantities that need to be evaluated for a given choice of cosmological paramteres. 
+
+This includes, the \mbox{\hyperlink{struct_class___cosmology___struct}{Class\+\_\+\+Cosmology\+\_\+\+Struct}} (initialized in \mbox{\hyperlink{cosmology_8c}{cosmology.\+c}}), and \mbox{\hyperlink{struct_line}{Line}} Structure (initialized in \mbox{\hyperlink{line__ingredients_8c}{line\+\_\+ingredients.\+c}}). 
+
+\doxysubsubsection{Field Documentation}
+\mbox{\Hypertarget{struct_cosmology_af8e7e270adf2573ca16f2037d580cb5b}\label{struct_cosmology_af8e7e270adf2573ca16f2037d580cb5b}} 
+\index{Cosmology@{Cosmology}!ccs@{ccs}}
+\index{ccs@{ccs}!Cosmology@{Cosmology}}
+\doxyparagraph{\texorpdfstring{ccs}{ccs}}
[…]
