+++ docs/latex/d1/dd5/ps__line__hm_8c.tex
@@ -1,0 +1,459 @@
+\hypertarget{ps__line__hm_8c}{}\doxysubsection{ps\+\_\+line\+\_\+hm.\+c File Reference}
+\label{ps__line__hm_8c}\index{ps\_line\_hm.c@{ps\_line\_hm.c}}
+
+
+Documented halo-\/model computation of line power spectrum, including clustering and stochastic contributions beyond Poisson limit.  
+
+
+{\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
+Include dependency graph for ps\+\_\+line\+\_\+hm.\+c\+:
+\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=350pt]{df/d3e/ps__line__hm_8c__incl}
+\end{center}
+\end{figure}
+\doxysubsubsection*{Functions}
+\begin{DoxyCompactItemize}
+\item 
+double \mbox{\hyperlink{ps__line__hm_8c_ab3acf266df779fd7b847b2ca641efa01}{PS\+\_\+line\+\_\+\+HM}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, double M\+\_\+min, long mode\+\_\+mf, long line\+\_\+type, int line\+\_\+id)
+\begin{DoxyCompactList}\small\item\em Compute the clustering contribution to the line power spectrum using halo-\/model. \end{DoxyCompactList}\item 
+double \mbox{\hyperlink{ps__line__hm_8c_ab3d6db65cacf00838ffc77e39ec93656}{PS\+\_\+shot\+\_\+\+HM}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double k, double z, double M\+\_\+min, double $\ast$input, long mode\+\_\+mf, long line\+\_\+type)
+\begin{DoxyCompactList}\small\item\em Compute the shot noise contributions, including corrections beyond poisson limit (see 1706.\+08738 for more details) If nfw=1, the dependance of the power spectrum on the halo profile is neglected. \end{DoxyCompactList}\item 
+static int \mbox{\hyperlink{ps__line__hm_8c_a122b3706f795f51c64ab8a9dd40379e3}{mhmc\+\_\+integ}} (const int $\ast$ndim, const cubareal x\mbox{[}$\,$\mbox{]}, const int $\ast$ncomp, cubareal ff\mbox{[}$\,$\mbox{]}, void $\ast$p)
+\begin{DoxyCompactList}\small\item\em The integrand function passed passed to Cuhre integration routine of CUBA library to compute the corrections to mass integration. \end{DoxyCompactList}\item 
+void \mbox{\hyperlink{ps__line__hm_8c_a3a71ccb997cf820c19eb64cedbf69a10}{mhmc}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double z, long mode\+\_\+mf, double $\ast$result)
+\begin{DoxyCompactList}\small\item\em Compute the corrections to mass integration of HM matter power spectrum. \end{DoxyCompactList}\item 
+static int \mbox{\hyperlink{ps__line__hm_8c_ad72b2846b2e235de535da531f2ee30da}{HM\+\_\+1h2h\+\_\+integ}} (const int $\ast$ndim, const cubareal x\mbox{[}$\,$\mbox{]}, const int $\ast$ncomp, cubareal ff\mbox{[}$\,$\mbox{]}, void $\ast$p)
[…]
