+++ docs/latex/d9/dd6/_r_e_a_d_m_e_8md.tex
@@ -1,0 +1,2 @@
+\hypertarget{_r_e_a_d_m_e_8md}{}\doxysubsection{README.\+md File Reference}
+\label{_r_e_a_d_m_e_8md}\index{README.md@{README.md}}
