+++ Include/ps_line_hm.h
@@ -1,5 +1,5 @@
-#ifndef LINE_PS_HM_H
-#define LINE_PS_HM_H
+#ifndef _LINE_PS_HM_H_
+#define _LINE_PS_HM_H_
 
 // double PS_line_HM_RSD(struct Cosmology *Cx, double k, double mu, double z, size_t line_id, struct Cosmology *Cx_ref);
 double PS_line_HM(struct Cosmology *Cx, double k, double z, double M_min, long mode_mf,  long line_type, int line_id);
