+++ Include/ir_res.h
@@ -1,6 +1,6 @@
 
-#ifndef ir_RES_H
-#define ir_RES_H
+#ifndef _IR_RES_H_
+#define _IR_RES_H_
 
 double pm_nowiggle_dst(struct Cosmology *Cx, double k,double z, int mode);
 double pm_nowiggle_gfilter(struct Cosmology *Cx, double k, double z, int mode);
