+++ Include/ps_halo_resps.h
@@ -1,12 +1,0 @@
-
-double Pk_rspace(double k, double z, double nbar, struct Cosmology *Cx);
-double Pk_zspace(double k, double mu, double z, double nbar, struct Cosmology *Cx);
-
-
-double resp_Pk_rspace(double k, double z, double R, double nbar, struct Cosmology *Cx);
-double resp_Pk_zspace(double k, double mu, double z, double R, double nbar, struct Cosmology *Cx);
-double resp_Delta0_rspace(double k, double z, double nbar, struct Cosmology *Cx);
-double resp_phi_rspace(double k, double z, double nbar, struct Cosmology *Cx);
-double resp_Delta0_zspace(double k, double mu, double z, double nbar, struct Cosmology *Cx);
-double resp_Delta2_zspace(double k, double mu, double z, double nbar, struct Cosmology *Cx);
-double resp_phi_zspace(double k, double mu, double z, double nbar, struct Cosmology *Cx);
