+++ docs/latex/df/dcb/header_8h.tex
@@ -1,536 +1,0 @@
-\hypertarget{header_8h}{}\doxysubsection{header.\+h File Reference}
-\label{header_8h}\index{header.h@{header.h}}
-{\ttfamily \#include $<$time.\+h$>$}\newline
-{\ttfamily \#include $<$unistd.\+h$>$}\newline
-{\ttfamily \#include $<$stdlib.\+h$>$}\newline
-{\ttfamily \#include $<$stdio.\+h$>$}\newline
-{\ttfamily \#include $<$math.\+h$>$}\newline
-{\ttfamily \#include $<$float.\+h$>$}\newline
-{\ttfamily \#include $<$string.\+h$>$}\newline
-{\ttfamily \#include $<$omp.\+h$>$}\newline
-{\ttfamily \#include $<$mpi.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+errno.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+spline.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+interp2d.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+spline2d.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+sf\+\_\+bessel.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+sf\+\_\+legendre.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+integration.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+matrix.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+linalg.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+blas.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+monte.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+monte\+\_\+vegas.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+odeiv2.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+roots.\+h$>$}\newline
-{\ttfamily \#include $<$gsl/gsl\+\_\+sf\+\_\+expint.\+h$>$}\newline
-{\ttfamily \#include $<$ctype.\+h$>$}\newline
-{\ttfamily \#include \char`\"{}cuba.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}cubature.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}global\+\_\+structs.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}utilities.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}cosmology.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}survey\+\_\+specs.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}line\+\_\+ingredients.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}wnw\+\_\+split.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}ir\+\_\+res.\+h\char`\"{}}\newline
-{\ttfamily \#include \char`\"{}ps\+\_\+halo\+\_\+1loop.\+h\char`\"{}}\newline
[…]
