+++ docs/latex/dir_74389ed8173ad57b461b9d623a1f3867.tex
@@ -1,0 +1,32 @@
+\hypertarget{dir_74389ed8173ad57b461b9d623a1f3867}{}\doxysubsection{Source Directory Reference}
+\label{dir_74389ed8173ad57b461b9d623a1f3867}\index{Source Directory Reference@{Source Directory Reference}}
+Directory dependency graph for Source\+:\nopagebreak
+\begin{figure}[H]
+\begin{center}
+\leavevmode
+\includegraphics[width=134pt]{dir_74389ed8173ad57b461b9d623a1f3867_dep}
+\end{center}
+\end{figure}
+\doxysubsubsection*{Files}
+\begin{DoxyCompactItemize}
+\item 
+file \mbox{\hyperlink{cosmology_8c}{cosmology.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented cosmology module. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{ir__res_8c}{ir\+\_\+res.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented IR\+\_\+res module. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{line__ingredients_8c}{line\+\_\+ingredients.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented line\+\_\+ingredients module. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{main_8c}{main.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented main module, including functions to initilize and cleanup the cosmology structure and examples of calls to functions in other modules to compute the line clustering and shot power spectrum. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{ps__halo__1loop_8c}{ps\+\_\+halo\+\_\+1loop.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented real-\/space, direct integration computation of 1loop contributions of the halo/galaxy power spectrum See ar\+Xiv\+:2010.\+14523 for explicit expressions. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{ps__line__hm_8c}{ps\+\_\+line\+\_\+hm.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented halo-\/model computation of line power spectrum, including clustering and stochastic contributions beyond Poisson limit. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{ps__line__pt_8c}{ps\+\_\+line\+\_\+pt.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented computation of Poisson shot noise and tree-\/level line power spectrum in real and redshift-\/space. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{survey__specs_8c}{survey\+\_\+specs.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented computation of some survey-\/related functions. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{utilities_8c}{utilities.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented basic utility functions used by other modules of the code. \end{DoxyCompactList}\item 
+file \mbox{\hyperlink{wnw__split_8c}{wnw\+\_\+split.\+c}}
+\begin{DoxyCompactList}\small\item\em Documented wiggle-\/nowiggle split based on 3d Gaussian filter in linear k, and using the Eisentein-\/\+Hu wiggle-\/no wiggle template ar\+Xiv\+:astro-\/ph/9709112. \end{DoxyCompactList}\end{DoxyCompactItemize}
