+++ docs/latex/df/dcb/header_8h_source.tex
@@ -1,0 +1,181 @@
+\hypertarget{header_8h_source}{}\doxysubsection{header.\+h}
+\label{header_8h_source}\index{header.h@{header.h}}
+\mbox{\hyperlink{header_8h}{Go to the documentation of this file.}}
+\begin{DoxyCode}{0}
+\DoxyCodeLine{1 }
+\DoxyCodeLine{6 \textcolor{preprocessor}{\#ifndef HEADER\_H}}
+\DoxyCodeLine{7 \textcolor{preprocessor}{\#define HEADER\_H}}
+\DoxyCodeLine{8 }
+\DoxyCodeLine{9 \textcolor{preprocessor}{\#define \_GNU\_SOURCE}}
+\DoxyCodeLine{10 }
+\DoxyCodeLine{11 \textcolor{preprocessor}{\#include <time.h>}}
+\DoxyCodeLine{12 \textcolor{preprocessor}{\#include <unistd.h>}}
+\DoxyCodeLine{13 \textcolor{preprocessor}{\#include <stdlib.h>}}
+\DoxyCodeLine{14 \textcolor{preprocessor}{\#include <stdio.h>}}
+\DoxyCodeLine{15 \textcolor{preprocessor}{\#include <math.h>}}
+\DoxyCodeLine{16 \textcolor{preprocessor}{\#include <float.h>}}
+\DoxyCodeLine{17 \textcolor{preprocessor}{\#include <string.h>}}
+\DoxyCodeLine{18 \textcolor{preprocessor}{\#include <omp.h>}}
+\DoxyCodeLine{19 \textcolor{preprocessor}{\#include <mpi.h>}}
+\DoxyCodeLine{20 \textcolor{preprocessor}{\#include <gsl/gsl\_errno.h>}}
+\DoxyCodeLine{21 \textcolor{preprocessor}{\#include <gsl/gsl\_spline.h>}}
+\DoxyCodeLine{22 \textcolor{preprocessor}{\#include <gsl/gsl\_interp2d.h>}}
+\DoxyCodeLine{23 \textcolor{preprocessor}{\#include <gsl/gsl\_spline2d.h>}}
+\DoxyCodeLine{24 \textcolor{preprocessor}{\#include <gsl/gsl\_sf\_bessel.h>}}
+\DoxyCodeLine{25 \textcolor{preprocessor}{\#include <gsl/gsl\_sf\_legendre.h>}}
+\DoxyCodeLine{26 \textcolor{preprocessor}{\#include <gsl/gsl\_integration.h>}}
+\DoxyCodeLine{27 \textcolor{preprocessor}{\#include <gsl/gsl\_matrix.h>}}
+\DoxyCodeLine{28 \textcolor{preprocessor}{\#include <gsl/gsl\_linalg.h>}}
+\DoxyCodeLine{29 \textcolor{preprocessor}{\#include <gsl/gsl\_blas.h>}}
+\DoxyCodeLine{30 \textcolor{preprocessor}{\#include <gsl/gsl\_monte.h>}}
+\DoxyCodeLine{31 \textcolor{preprocessor}{\#include <gsl/gsl\_monte\_vegas.h>}}
+\DoxyCodeLine{32 \textcolor{preprocessor}{\#include <gsl/gsl\_odeiv2.h>}    }
+\DoxyCodeLine{33 \textcolor{preprocessor}{\#include <gsl/gsl\_roots.h>}   }
+\DoxyCodeLine{34 \textcolor{preprocessor}{\#include <gsl/gsl\_sf\_expint.h>}}
+\DoxyCodeLine{35 \textcolor{preprocessor}{\#include <ctype.h>}}
[…]

[…]
