### 2022-02-17 amoradinejad/limHaloPT#2 @ziotom78
Circular reference in makefile
There is a circular reference in the file `Source/makefile`:

```
line_ingredients.o: line_ingredients.o ../Include/line_ingredients.h
        mpicc -g -c  -fopenmp line_ingredients.c -o line_ingredients.o $(INCLUDES)
```

but it should read

```
#                                    ↓   Modify the extension here
line_ingredients.o: line_ingredients.c ../Include/line_ingredients.h
        mpicc -g -c  -fopenmp line_ingredients.c -o line_ingredients.o $(INCLUDES)
```

(This issue has been opened as part of the [review of the JOSS paper](https://github.com/openjournals/joss-reviews/issues/4173).)
