### 2022-02-18 df386a2
platform, compiler and OS added.
