### 2022-02-18 amoradinejad/limHaloPT#2 @amoradinejad
Circular reference corrected.
