+++ Class/include/background.h
@@ -1,636 +1,0 @@
-/** @file background.h Documented includes for background module */
-
-#ifndef __BACKGROUND__
-#define __BACKGROUND__
-
-#include "common.h"
-#include "quadrature.h"
-#include "growTable.h"
-#include "arrays.h"
-#include "dei_rkck.h"
-#include "parser.h"
-
-/** list of possible types of spatial curvature */
-
-enum spatial_curvature {flat,open,closed};
-
-/** list of possible parametrisations of the DE equation of state */
-
-enum equation_of_state {CLP,EDE};
-
-
-/** list of possible parametrizations of the varying fundamental constants */
-
-enum varconst_dependence {varconst_none,varconst_instant};
-
-/** list of formats for the vector of background quantities */
-
-enum vecback_format {short_info, normal_info, long_info};
-
-/** list of interpolation methods: search location in table either
-    by bisection (inter_normal), or step by step starting from given
-    index (inter_closeby) */
-
-enum interpolation_method {inter_normal, inter_closeby};
-
-/**
- * background structure containing all the background information that
- * other modules need to know.
- *
- * Once initialized by the backgound_init(), contains all necessary
- * information on the background evolution (except thermodynamics),
- * and in particular, a table of all background quantities as a
- * function of time and scale factor, used for interpolation in other
- * modules.
- */
-
-struct background
-{
-  /** @name - input parameters initialized by user in input module
-   *  (all other quantities are computed in this module, given these parameters
-   *   and the content of the 'precision' structure)
-   *
-   * The background cosmological parameters listed here form a parameter
-   * basis which is directly usable by the background module. Nothing
-   * prevents from defining the input cosmological parameters
-   * differently, and to pre-process them into this format, using the input
-   * module (this might require iterative calls of background_init()
-   * e.g. for dark energy or decaying dark matter). */
-
-  //@{
-
-  double H0; /**< \f$ H_0 \f$: Hubble parameter (in fact, [\f$H_0/c\f$]) in \f$ Mpc^{-1} \f$ */
-  double h;  /**< reduced Hubble parameter */
-
-  double Omega0_g; /**< \f$ \Omega_{0 \gamma} \f$: photons */
-  double T_cmb;    /**< \f$ T_{cmb} \f$: current CMB temperature in Kelvins */
-
-  double Omega0_b; /**< \f$ \Omega_{0 b} \f$: baryons */
-
-  double Omega0_ur; /**< \f$ \Omega_{0 \nu r} \f$: ultra-relativistic neutrinos */
-
-  double Omega0_cdm;      /**< \f$ \Omega_{0 cdm} \f$: cold dark matter */
-
-  double Omega0_idr; /**< \f$ \Omega_{0 idr} \f$: interacting dark radiation */
-  double T_idr;      /**< \f$ T_{idr} \f$: current temperature of interacting dark radiation in Kelvins */
-
-  double Omega0_idm_dr; /**< \f$ \Omega_{0 idm_dr} \f$: dark matter interacting with dark radiation */
-
-  double Omega0_dcdmdr;   /**< \f$ \Omega_{0 dcdm}+\Omega_{0 dr} \f$: decaying cold dark matter (dcdm) decaying to dark radiation (dr) */
-  double Omega_ini_dcdm;  /**< \f$ \Omega_{ini,dcdm} \f$: rescaled initial value for dcdm density (see 1407.2418 for definitions) */
-  double Gamma_dcdm;      /**< \f$ \Gamma_{dcdm} \f$: decay constant for decaying cold dark matter */
-  double tau_dcdm;
-
-  int N_ncdm;                            /**< Number of distinguishable ncdm species */
-  /* the following parameters help to define tabulated ncdm p-s-d passed in file */
[…]

[…]
