+++ Class/include/common.h
@@ -1,0 +1,407 @@
+/** @file common.h Generic libraries, parameters and functions used in the whole code. */
+
+#include "stdio.h"
+#include "stdlib.h"
+#include "math.h"
+#include "string.h"
+#include "float.h"
+#include "svnversion.h"
+#include <stdarg.h>
+
+#ifdef _OPENMP
+#include "omp.h"
+#endif
+
+#ifndef __COMMON__
+#define __COMMON__
+
+#define _VERSION_ "v3.1.1"
+
+/* @cond INCLUDE_WITH_DOXYGEN */
+
+#define _TRUE_ 1 /**< integer associated to true statement */
+#define _FALSE_ 0 /**< integer associated to false statement */
+
+#define _SUCCESS_ 0 /**< integer returned after successful call of a function */
+#define _FAILURE_ 1 /**< integer returned after failure in a function */
+
+#define _ERRORMSGSIZE_ 2048 /**< generic error messages are cut beyond this number of characters */
+typedef char ErrorMsg[_ERRORMSGSIZE_]; /**< Generic error messages (there is such a field in each structure) */
+
+#define _FILENAMESIZE_ 256 /**< size of the string read in each line of the file (extra characters not taken into account) */
+typedef char FileName[_FILENAMESIZE_];
+
+#define _SUFFIXNAMESIZE_ 4 /**< maximum size of the short string appended to file names to account for initial conditions, etc. */
+
+#define _PI_ 3.1415926535897932384626433832795e0 /**< The number pi */
+
+#define _PIHALF_ 1.57079632679489661923132169164e0 /**< pi divided by 2 */
+
+#define _TWOPI_ 6.283185307179586476925286766559e0 /**< 2 times pi */
+
+#define _SQRT2_ 1.41421356237309504880168872421e0 /** < square root of 2. */
+
+#define _SQRT6_ 2.4494897427831780981972840747059e0 /**< square root of 6. */
+
+#define _SQRT_PI_ 1.77245385090551602729816748334e0 /**< square root of pi. */
+
+#define _E_ 2.718281828459045235360287471352662497757247093699959574966967627724076630353547594571382178525166427427466391932003059921817413596629043572900334295260595630738132328627943490763233829880753195251019011573834187930702154089149934884167509244761460668082264800168477411853742345442437107539077744992069551702761838606261331384583000752044933826560297606737113200709328709127443747047230696977209310141692836819025515108657463772111252389784425056953696 /**< exponential of one */
+
+#define _MAX_IT_ 10000/**< default maximum number of iterations in conditional loops (to avoid infinite loops) */
+
+#define _QUADRATURE_MAX_ 250 /**< maximum allowed number of abssices in quadrature integral estimation */
+
+#define _QUADRATURE_MAX_BG_ 800 /**< maximum allowed number of abssices in quadrature integral estimation */
+
+#define _TOLVAR_ 100. /**< The minimum allowed variation is the machine precision times this number */
+
+#define _HUGE_ 1.e99
+
+#define _EPSILON_ 1.e-10
+
+#define _OUTPUTPRECISION_ 12 /**< Number of significant digits in some output files */
+
+#define _COLUMNWIDTH_ 24 /**< Must be at least _OUTPUTPRECISION_+8 for guaranteed fixed width columns */
+
+#define _MAXTITLESTRINGLENGTH_ 8000 /**< Maximum number of characters in title strings */
+
+#define _DELIMITER_ "\t" /**< character used for delimiting titles in the title strings */
+
+#ifndef __CLASSDIR__
+#define __CLASSDIR__ "." /**< The directory of CLASS. This is set to the absolute path to the CLASS directory so this is just a failsafe. */
+#endif
+
[…]
