+++ Class/include/perturbations.h
@@ -1,0 +1,951 @@
+/** @file perturbations.h Documented includes for perturbation module */
+
+#ifndef __PERTURBATIONS__
+#define __PERTURBATIONS__
+
+#include "thermodynamics.h"
+
+#define _scalars_ ((ppt->has_scalars == _TRUE_) && (index_md == ppt->index_md_scalars))
+#define _vectors_ ((ppt->has_vectors == _TRUE_) && (index_md == ppt->index_md_vectors))
+#define _tensors_ ((ppt->has_tensors == _TRUE_) && (index_md == ppt->index_md_tensors))
+
+#define _set_source_(index) ppt->sources[index_md][index_ic * ppt->tp_size[index_md] + index][index_tau * ppt->k_size[index_md] + index_k]
+
+/**
+ * flags for various approximation schemes
+ * (tca = tight-coupling approximation,
+ *  rsa = radiation streaming approximation,
+ *  ufa = massless neutrinos / ultra-relativistic relics fluid approximation)
+ *
+ * CAUTION: must be listed below in chronological order, and cannot be
+ * reversible. When integrating equations for a given mode, it is only
+ * possible to switch from left to right in the lists below.
+ */
+
+//@{
+
+enum tca_flags {tca_on, tca_off};
+enum rsa_flags {rsa_off, rsa_on};
+enum tca_idm_dr_flags {tca_idm_dr_on, tca_idm_dr_off};
+enum rsa_idr_flags {rsa_idr_off, rsa_idr_on};
+enum ufa_flags {ufa_off, ufa_on};
+enum ncdmfa_flags {ncdmfa_off, ncdmfa_on};
+
+//@}
+
+/**
+ * labels for the way in which each approximation scheme is implemented
+ */
+
+//@{
+
+enum tca_method {first_order_MB,first_order_CAMB,first_order_CLASS,second_order_CRS,second_order_CLASS,compromise_CLASS};
+enum rsa_method {rsa_null,rsa_MD,rsa_MD_with_reio,rsa_none};
+enum idr_method {idr_free_streaming,idr_fluid}; /* for the idm-idr case */
+enum rsa_idr_method {rsa_idr_none,rsa_idr_MD};  /* for the idm-idr case */
+enum ufa_method {ufa_mb,ufa_hu,ufa_CLASS,ufa_none};
+enum ncdmfa_method {ncdmfa_mb,ncdmfa_hu,ncdmfa_CLASS,ncdmfa_none};
+enum tensor_methods {tm_photons_only,tm_massless_approximation,tm_exact};
+
+//@}
+
+/**
+ * List of coded gauges. More gauges can in principle be defined.
+ */
+
+//@{
+
+enum possible_gauges {
+  newtonian, /**< newtonian (or longitudinal) gauge */
+  synchronous /**< synchronous gauge with \f$ \theta_{cdm} = 0 \f$ by convention */
+};
+
+//@}
+
+//@{
+
+/**
+ * maximum number and types of selection function (for bins of matter density or cosmic shear)
+ */
+#define _SELECTION_NUM_MAX_ 100
+enum selection_type {gaussian,tophat,dirac};
+
+//@}
+
+
+//@{
+
+/**
+ * maximum number of k-values for perturbation output
+ */
+#define _MAX_NUMBER_OF_K_FILES_ 30
+
+//@}
+
+
+
+/**
+ * Structure containing everything about perturbations that other
+ * modules need to know, in particular tabled values of the source
+ * functions \f$ S(k, \tau) \f$ for all requested modes
+ * (scalar/vector/tensor), initial conditions, types (temperature,
+ * E-polarization, B-polarization, lensing potential, etc), multipole
+ * l and wavenumber k.
+ *
+ */
+
+struct perturbations
+{
[…]

[…]
