### 2022-02-18 2bc7c7a
README files edited to reflect compiling of class
