### 2022-02-18 d36bd6e
typo in README corrected
