### 2022-03-01 cd7be5b
readme edited to correctly give the subdirectory from which the code has to be compiled and run to fix #8.
