### 2022-03-02 amoradinejad/limHaloPT#10 @abhimaniyar
Example runs for the code
I installed the code following the instructions in the README file. In order to check if the code runs well, I did `./limHaloPT` as suggested in the README file. It seemed like the code was running, however, the run takes a while to complete. I realized that this was happening because of the `size_t ninterp = 150` parameter in the `main.c` file which gives number of points for redshift interpolation. In order to decrease the run time, I changed this parameter to `size_t ninterp = 5` and `nlines = 1` (just for 1 line) and ran the code. This is what I get:

```
$ ./limHaloPT
hm 131 0 0.000000e+00 1.971505e+02 1.604636e+06 1.536095e+02 -6.258736e+01
hm 131 1 2.600000e+00 1.594386e+03 5.543890e+08 3.445934e+03 1.644612e+03
hm 131 2 5.200000e+00 1.088735e+03 1.947773e+08 3.798287e+03 6.804291e+03
hm 131 3 7.800000e+00 6.936088e+02 3.475170e+07 3.116919e+03 8.353733e+03
hm 131 4 1.040000e+01 3.722656e+02 1.045218e+07 2.217445e+03 9.013637e+03
for line 131 mom1, mom2, b1 and b2 interpolators initialized
Reference Cosmology initialized
Elapsed: 107.390372 seconds for ref cosmology
Segmentation fault: 11
```
No output file is produced. I am not sure if this is due to the change in `size_t ninterp` parameter.
I think it would be great if a simpler sample run is provided to check the installation of the code which takes a reasonable time. Also, it would be good to mention what is the expected output of the sample run.
