### 2022-03-02 9619c8a
To fix #10, several comments added to main.c and ps_line_hm.c, related to ninterp, output file paths. Output directory created.
