### 2022-03-02 365a3a0
readme of the doxygen doc updated
