### 2022-03-15 amoradinejad/limHaloPT#11 @ziotom78
Lack of examples in the documentation
It is not apparent to me where to look for examples… Both the README and the documentation just tell how to compile the code and run it, but there is no plot/script/whatever showing the user how to employ the program to do some simple analysis.

Consider for instance [QuadGK.jl](https://github.com/JuliaMath/QuadGK.jl), a Julia package to compute 1D numerical integrals. The README shows how to call its functions at the very beginning of the text, with a simple example (integral of `exp(-x²)`).

Another example is [Zygote.jl](https://github.com/FluxML/Zygote.jl), again a Julia package, where the README starts with a simple example showing how the package can differentiate arbitrary functions.

In both cases the user has immediately an idea of how the code should be used, the kind of expected input, and the format used for the output.
