### 2022-03-15 #4173 @ziotom78
@amoradinejad, I have just posted another issue on your repository ([#11](https://github.com/amoradinejad/limHaloPT/issues/11)) to add examples to the README.

I have a general comment. The lack of examples in the README misses one of the requirements for a paper to be accepted by JOSS, as shown in the checklist that is shown above in this page and stated here («Documentation» / «Example usage»):

https://joss.readthedocs.io/en/latest/review_checklist.html

I encourage you to read carefully this list of points. After a quick check, I fear that other points are not addressed (are there «automatic tests»? «community guidelines»?).

If you could have a check of the kind of stuff JOSS reviewers ask, the review process would probably speed up enormously.
