### 2022-03-15 #4173 @abhimaniyar
@amoradinejad <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_amoradinejad&d=DwMFaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=87YU7psLWOQYeS5uKAEh4WOPw7Hvp2FXY9wC5FBHIa8Jh6j1Di48E5mPw3j-IrC2&s=MZ6DcO_fSJvwCDNGYqOgnqusAEaSg6brS2kUom3co80&e=>, I have just posted an issue (#12 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_amoradinejad_limHaloPT_issues_11&d=DwMFaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=87YU7psLWOQYeS5uKAEh4WOPw7Hvp2FXY9wC5FBHIa8Jh6j1Di48E5mPw3j-IrC2&s=JjPOIZE1Lm8mKFS2Hn10jdTkOpKY73bCu3RvCgGBNf0&e=>) following up on my previous issue (#10 <https://github.com/amoradinejad/limHaloPT/issues/10#issue-1156019803>) to expand your documentation to include more clarity on the sample examples for the users. Kindly update the README accordingly.
Thank you.

> On 15 Mar 2022, at 08:02, Maurizio Tomasi ***@***.***> wrote:
>
>
> @amoradinejad <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_amoradinejad&d=DwMFaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=87YU7psLWOQYeS5uKAEh4WOPw7Hvp2FXY9wC5FBHIa8Jh6j1Di48E5mPw3j-IrC2&s=MZ6DcO_fSJvwCDNGYqOgnqusAEaSg6brS2kUom3co80&e=>, I have just posted another issue on your repository (#11 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_amoradinejad_limHaloPT_issues_11&d=DwMFaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=87YU7psLWOQYeS5uKAEh4WOPw7Hvp2FXY9wC5FBHIa8Jh6j1Di48E5mPw3j-IrC2&s=JjPOIZE1Lm8mKFS2Hn10jdTkOpKY73bCu3RvCgGBNf0&e=>) to add examples to the README.
>
> I have a general comment. The lack of examples in the README misses one of the requirements for a paper to be accepted by JOSS, as shown in the checklist that is shown above in this page and stated here («Documentation» / «Example usage»):
>
> https://joss.readthedocs.io/en/latest/review_checklist.html <https://urldefense.proofpoint.com/v2/url?u=https-3A__joss.readthedocs.io_en_latest_review-5Fchecklist.html&d=DwMFaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=87YU7psLWOQYeS5uKAEh4WOPw7Hvp2FXY9wC5FBHIa8Jh6j1Di48E5mPw3j-IrC2&s=2MAZ1h2SyiVmgfe_s6uKyKad2ny3CLQuhg1aOBzH2v8&e=>
> I encourage you to read carefully this list of points. After a quick check, I fear that other points are not addressed (are there «automatic tests»? «community guidelines»?).
>
> If you could have a check of the kind of stuff JOSS reviewers ask, the review process would probably speed up enormously.
>
