### 2022-03-15 amoradinejad/limHaloPT#10 @amoradinejad
I have previously addressed the request of the referee in giving more information about a simple sample run by adding comments in main.c file which includes example calls to useful functions of the package.  It seems I have not correctly responded to this thread here, although the changes were made to main.c file and in the comment of the commit I have made;  "To fix https://github.com/amoradinejad/limHaloPT/issues/10, several comments added to main.c and ps_line_hm.c, related to ninterp, output file paths. Output directory created."

Please let me know if the requests are addressed.
