+++ docs/latex/dd/d24/survey__specs_8c.tex
@@ -6,7 +6,8 @@ Documented computation of some survey-\/related functions.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for survey\+\_\+specs.\+c\+:\nopagebreak
+Include dependency graph for survey\+\_\+specs.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
