+++ docs/latex/d7/d78/wnw__split_8c.tex
@@ -6,7 +6,8 @@ Documented wiggle-\/nowiggle split based on 3d Gaussian filter in linear k, and 
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for wnw\+\_\+split.\+c\+:\nopagebreak
+Include dependency graph for wnw\+\_\+split.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
