+++ docs/latex/d8/d43/ps__line__pt_8c.tex
@@ -6,7 +6,8 @@ Documented computation of Poisson shot noise and tree-\/level line power spectru
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:\nopagebreak
+Include dependency graph for ps\+\_\+line\+\_\+pt.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
