+++ docs/latex/d0/d29/main_8c.tex
@@ -6,7 +6,8 @@ Documented main module, including functions to initilize and cleanup the cosmolo
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for main.\+c\+:\nopagebreak
+Include dependency graph for main.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -112,11 +113,20 @@ Here is the caller graph for this function\+:\nopagebreak
 
 Define which lines to compute, and how many points to have for z-\/interpolation
 
+Note\+: The main line properties used in the calculation of the line power spectrum, which includes first and second moments of line luminosity (weighted by halo mass function) and luminosity-\/weighted linear and quadratic line biases are computed in function \mbox{\hyperlink{line__ingredients_8c_aa4696edc9829e9a3c2c1012a3e9d9dbe}{Line\+\_\+alloc\+\_\+init()}}. In that function, these quantities are calculate over an array of redshift values and then an interpolator for these quantities is initialized. The parameter \char`\"{}ninterp\char`\"{} below, sets the number of interpolation points ~\newline
+ for this interpolator. If you set it too low, the redshift-\/dependance of line properties would not be accurate. In the tests done during code developement, ninterp = 150 give results at the precision needed. However large values of ninterp results in longer run time.
+
+For testing installation and running of the code, you can set lower value of ninterp, to make the code run faster. But remember in your actuall runs, set ninterp to larger numbers.
+
+If you do not need to compute the line properties for all the 7 lines below, you can set nlines to the number of lines you are interested, change the arrays of line\mbox{[}nlines\mbox{]} and JJ\mbox{[}nlines\mbox{]} to correspond to the lines you need to calculate.
+
 Initialize the cosmology structure, which includes CLASS cosmology and \mbox{\hyperlink{struct_line}{Line}} structures
 
 \DoxyHorRuler{0}
  \hypertarget{main.c_autotoc_md1}{}\doxyparagraph{Depending on what quantities ou want to compute, you need to modify this part of the main() function.}\label{main.c_autotoc_md1}
 Set the k and z arrays
+
+For testing the code, you can reduce the range of k, and z and also the number of points for each, nk, and nz.
 
 Compute the line clustering signal using halo model
 
