+++ docs/latex/d7/d10/line__ingredients_8c.tex
@@ -6,7 +6,8 @@ Documented line\+\_\+ingredients module.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for line\+\_\+ingredients.\+c\+:\nopagebreak
+Include dependency graph for line\+\_\+ingredients.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -572,7 +573,7 @@ Read in the measured mass function of Hidden-\/valey sims and build an interpola
 the interpolated measured halo mass function ~\newline
  M in unit of M\+\_\+sun and HMF in unit of \#-\/of-\/halos/\+Mpc$^\wedge$3/\+M\+\_\+sun 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+not to be used at z other than 2. This function is for testing purposes only. We test the theoretical predictions at z=2 if ST MF or measured MF are used.Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
