+++ docs/latex/d1/d48/ir__res_8c.tex
@@ -6,7 +6,8 @@ Documented IR\+\_\+res module.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ir\+\_\+res.\+c\+:\nopagebreak
+Include dependency graph for ir\+\_\+res.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
