+++ docs/latex/d8/d2a/ps__halo__1loop_8c.tex
@@ -6,7 +6,8 @@ Documented real-\/space, direct integration computation of 1loop contributions o
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+halo\+\_\+1loop.\+c\+:\nopagebreak
+Include dependency graph for ps\+\_\+halo\+\_\+1loop.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
