+++ docs/latex/d7/db6/utilities_8c.tex
@@ -6,7 +6,8 @@ Documented basic utility functions used by other modules of the code.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for utilities.\+c\+:\nopagebreak
+Include dependency graph for utilities.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
