### 2022-03-15 amoradinejad/limHaloPT#10 @abhimaniyar
According to the JOSS requirements (here: https://joss.readthedocs.io/en/latest/review_checklist.html <https://urldefense.proofpoint.com/v2/url?u=https-3A__joss.readthedocs.io_en_latest_review-5Fchecklist.html&d=DwMFaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=87YU7psLWOQYeS5uKAEh4WOPw7Hvp2FXY9wC5FBHIa8Jh6j1Di48E5mPw3j-IrC2&s=2MAZ1h2SyiVmgfe_s6uKyKad2ny3CLQuhg1aOBzH2v8&e=>), README should include examples and their details i.e. what is being run/what is the expected output/what are the things code can calculate and which functions to call for them etc.

So I think it would be helpful to add these details in the README section.

> On 15 Mar 2022, at 19:47, Azadeh Moradinezhad ***@***.***> wrote:
>
>
> I have previously addressed the request of the referee in giving more information about a simple sample run by adding comments in main.c file which includes example calls to useful functions of the package. It seems I have not correctly responded to this thread here, although the changes were made to main.c file and in the comment of the commit I have made; "To fix #10 <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_amoradinejad_limHaloPT_issues_10&d=DwMCaQ&c=slrrB7dE8n7gBJbeO0g-IQ&r=j7NTX35lzotSaEysgWTW31rDj0okOldBCLczNlYAI8w&m=roVnwTmW67Sr2xVeonDP5sxGCAqBQZh2V9MELCaOY-h1bEbgucWQ00J0eZ7Ta9Lk&s=5KV9ReVw7WZ9IYP-b6jRKfqkzCrpE1UTEmVurWgD0Nw&e=>, several comments added to main.c and ps_line_hm.c, related to ninterp, output file paths. Output directory created."
>
> Please let me know if the requests are addressed.
>
