### 2022-03-16 f924e2b
all remaining instances of declaration of gb removed except in main.c
