+++ Source/ps_line_hm.c
@@ -18,7 +18,6 @@
  */
 
 #include "header.h"
-struct globals gb;
 
 
 /**
