+++ Source/ps_line_pt.c
@@ -17,7 +17,6 @@
  */
 
 #include "header.h"
-struct globals gb;
 
 
 /**
