+++ Source/survey_specs.c
@@ -11,8 +11,6 @@
  */
 
 #include "header.h"
-struct globals gb;
-
 
 
 /**
