+++ Source/line_ingredients.c
@@ -30,7 +30,6 @@
 
 
 #include "header.h"
-struct globals gb;
 
 
 /**
