### 2022-04-03 fc65d9a
partially addressed #10, #11, #12. For the moment, additional section describing the structure of the code is also added. Furthermore, the main.c module is edited to add additional comments.
