+++ README.md
@@ -96,3 +96,5 @@ Copyright 2021 Azadeh Moradinezhad Dizgah.<br>
 limHaloPT is free software made available under the MIT License. For details see the [LICENSE](https://github.com/amoradinejad/limHaloPT/blob/d40a4a75188ae70f56ed76236d1fd9ee1aae312d/LICENSE) file.
 
 
+## Contributing
+To contribute, create a fork on github, make changes and commits, and submit a pull request on github. If you found any issues with the code, please get in touch by posting an issue on this github repository.
