### 2022-04-03 f005702
makefile edited to make it cleaner
