+++ Source/cubature.c
@@ -1,1355 +1,0 @@
-/* Adaptive multidimensional integration of a vector of integrands.
- *
- * Copyright (c) 2005-2010 Steven G. Johnson
- *
- * Portions (see comments) based on HIntLib (also distributed under
- * the GNU GPL, v2 or later), copyright (c) 2002-2005 Rudolf Schuerer.
- *     (http://www.cosy.sbg.ac.at/~rschuer/hintlib/)
- *
- * Portions (see comments) based on GNU GSL (also distributed under
- * the GNU GPL, v2 or later), copyright (c) 1996-2000 Brian Gough.
- *     (http://www.gnu.org/software/gsl/)
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- *
- */
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <math.h>
-#include <limits.h>
-#include <float.h>
-#include "header.h"
-
-/* Adaptive multidimensional integration on hypercubes (or, really,
-   hyper-rectangles) using cubature rules.
-
-   A cubature rule takes a function and a hypercube and evaluates
-   the function at a small number of points, returning an estimate
-   of the integral as well as an estimate of the error, and also
-   a suggested dimension of the hypercube to subdivide.
-
-   Given such a rule, the adaptive integration is simple:
-
-   1) Evaluate the cubature rule on the hypercube(s).
-      Stop if converged.
-
-   2) Pick the hypercube with the largest estimated error,
-      and divide it in two along the suggested dimension.
-
-   3) Goto (1).
-
- The basic algorithm is based on the adaptive cubature described in
- 
-     A. C. Genz and A. A. Malik, "An adaptive algorithm for numeric
-     integration over an N-dimensional rectangular region,"
-     J. Comput. Appl. Math. 6 (4), 295-302 (1980).
-
- and subsequently extended to integrating a vector of integrands in
-
-     J. Berntsen, T. O. Espelid, and A. Genz, "An adaptive algorithm
-     for the approximate calculation of multiple integrals,"
-     ACM Trans. Math. Soft. 17 (4), 437-451 (1991).
-
- Note, however, that we do not use any of code from the above authors
- (in part because their code is Fortran 77, but mostly because it is
- under the restrictive ACM copyright license).  I did make use of some
- GPL code from Rudolf Schuerer's HIntLib and from the GNU Scientific
[…]
