### 2022-04-03 e046b3a
makefile edited to make it cleaner
