### 2022-04-03 1aede82
moved makefile to main limHaloPT directory, improved makefile, and move initialize() and cleanup() to a seperate module
