+++ Include/setup_teardown.h
@@ -1,0 +1,12 @@
+
+#ifndef _SETUP_TEARDOWN_H_
+#define _SETUP_TEARDOWN_H_
+
+/** 
+ * Function declarations of main.c module
+ */
+void  initialize();
+void  cleanup();
+
+
+#endif
