### 2022-04-03 114da4f
automatic test added, still in progress
