+++ docs/latex/refman.tex
@@ -201,6 +201,7 @@
 \input{d8/d2a/ps__halo__1loop_8c}
 \input{d1/dd5/ps__line__hm_8c}
 \input{d8/d43/ps__line__pt_8c}
+\input{d2/db8/setup__teardown_8c}
 \input{dd/d24/survey__specs_8c}
 \input{d7/db6/utilities_8c}
 \input{d7/d78/wnw__split_8c}
