+++ docs/latex/d1/dd5/ps__line__hm_8c.tex
@@ -6,8 +6,7 @@ Documented halo-\/model computation of line power spectrum, including clustering
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+line\+\_\+hm.\+c\+:
-\nopagebreak
+Include dependency graph for ps\+\_\+line\+\_\+hm.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -33,11 +32,6 @@ static int \mbox{\hyperlink{ps__line__hm_8c_a4e1bbff7f992103c9b1fa8f5abc1f408}{b
 \begin{DoxyCompactList}\small\item\em The integrand function passed passed to Cuhre integration routine to compute large-\/scale limit of b22 (shot-\/noise contribution) \end{DoxyCompactList}\item 
 double \mbox{\hyperlink{ps__line__hm_8c_aebed249f1a0afb0c6f7bcadcd3f1f4ee}{b22\+\_\+ls}} (struct \mbox{\hyperlink{struct_cosmology}{Cosmology}} $\ast$Cx, double z)
 \begin{DoxyCompactList}\small\item\em Compute the large-\/scale limit of P\+\_\+b2b2 loop. \end{DoxyCompactList}\end{DoxyCompactItemize}
-\doxysubsubsection*{Variables}
-\begin{DoxyCompactItemize}
-\item 
-struct \mbox{\hyperlink{structglobals}{globals}} \mbox{\hyperlink{ps__line__hm_8c_a61c6674f2b410e126a682325ce5360c0}{gb}}
-\end{DoxyCompactItemize}
 
 
 \doxysubsubsection{Detailed Description}
@@ -432,12 +426,3 @@ Here is the caller graph for this function\+:\nopagebreak
 \includegraphics[width=229pt]{d1/dd5/ps__line__hm_8c_ab3d6db65cacf00838ffc77e39ec93656_icgraph}
 \end{center}
 \end{figure}
-
-
-\doxysubsubsection{Variable Documentation}
-\mbox{\Hypertarget{ps__line__hm_8c_a61c6674f2b410e126a682325ce5360c0}\label{ps__line__hm_8c_a61c6674f2b410e126a682325ce5360c0}} 
-\index{ps\_line\_hm.c@{ps\_line\_hm.c}!gb@{gb}}
-\index{gb@{gb}!ps\_line\_hm.c@{ps\_line\_hm.c}}
-\doxyparagraph{\texorpdfstring{gb}{gb}}
-{\footnotesize\ttfamily struct \mbox{\hyperlink{structglobals}{globals}} gb}
-
