+++ docs/latex/index.tex
@@ -14,7 +14,7 @@ This package is released together with the following publication, \href{https://
 The source code of this package is publically avaialbel on Git\+Hub at \href{https://github.com/amoradinejad/limHaloPT}{\texttt{ lim\+Halo\+PT}}. ~\newline
  ~\hypertarget{index_autotoc_md13}{}\doxysubsubsection{Dependencies}\label{index_autotoc_md13}
 The lim\+Halo\+PT package calls various functions from \href{https://github.com/lesgourg/class_public}{\texttt{ CLASS}} Boltzman solver, including the matter power spectrum and transfer functions, growth factor etc. Therefore, you need to first download and compile CLASS code, create a \char`\"{} CLASS/lib/\char`\"{} folder and place the \char`\"{}libclass.\+a\char`\"{} in that folder. Furtehrmore, the loop calculations are performed with direct numerical integration, using routines of \href{http://www.feynarts.de/cuba/}{\texttt{ CUBA}} library. Furthermore, the code heavily uses functions of \href{https://www.gnu.org/software/gsl/doc/html/}{\texttt{ GSL}} scientific library. Therfore, make sure that the two libraries are correctly linked to lim\+Halo\+PT by making necassary modifcations to the makefile (placed in Source directory) of lim\+Halo\+PT package. ~\newline
- ~\hypertarget{index_autotoc_md14}{}\doxysubsection{Compilation}\label{index_autotoc_md14}
+ ~\hypertarget{index_autotoc_md14}{}\doxysubsubsection{Compilation}\label{index_autotoc_md14}
 
 \begin{DoxyItemize}
 \item To compile, type\+: make ~\newline
@@ -22,7 +22,7 @@ The lim\+Halo\+PT package calls various functions from \href{https://github.com/
 \end{DoxyItemize}
 
 If you modified the code, you need to first do \char`\"{}make clean\char`\"{} before doing \char`\"{}make\char`\"{}. The entire lim\+Halo\+PT package was developed, compiled, and tested on Mac OS X, using gcc version 7.\+5.\+0 compiler. ~\newline
- ~\hypertarget{index_autotoc_md15}{}\doxysubsection{Basic usage}\label{index_autotoc_md15}
+ ~\hypertarget{index_autotoc_md15}{}\doxysubsubsection{Basic usage}\label{index_autotoc_md15}
 In the current version of lim\+Halo\+PT, to use the package, you need to modify the \mbox{\hyperlink{main_8c}{main.\+c}} module. After each modification, the package needs to be re-\/compiled before it can be run. In the future versions, you should be able to just set an .ini file to set which quantities to be computed and the values of the parmaeters to be used.
 
 Depending on what quantities you want to calculate, you should modify the \mbox{\hyperlink{main_8c_a0ddf1224851353fc92bfbff6f499fa97}{main()}} function in \mbox{\hyperlink{main_8c}{main.\+c}} module (as marked in the code). Before calling any function within \mbox{\hyperlink{main_8c}{main.\+c}}, you may want to also change default values for some other initialization steps. This is also marked in the code. Two example calls to functions that compute the clustering and shot noise contributions aree included in \mbox{\hyperlink{main_8c}{main.\+c}} module. We descibe one of them here. ~\newline
@@ -32,6 +32,6 @@ You can use this package freely, provided that in your publication you cite the 
 \href{https://opensource.org/licenses/MIT}{\texttt{ }} ~\newline
  Copyright 2021 Azadeh Moradinezhad Dizgah.~\newline
[…]
