+++ docs/latex/index.tex
@@ -18,7 +18,7 @@ The lim\+Halo\+PT package calls various functions from \href{https://github.com/
 
 \begin{DoxyItemize}
 \item To compile, type\+: make ~\newline
- ~
+
 \end{DoxyItemize}
 
 If you modified the code, you need to first do \char`\"{}make clean\char`\"{} before doing \char`\"{}make\char`\"{}. The entire lim\+Halo\+PT package was developed, compiled, and tested on Mac OS X, using gcc version 7.\+5.\+0 compiler. ~\newline
