+++ Source/main.c
@@ -14,79 +14,87 @@ struct globals gb;
 
 
 int main(int argc,char *argv[])
-{	
- 	
- 	/**
-	 * -----------------------------
-	 * If you want to change the values of cosmological parmaeters compared to the 
-	 * default values of the code, you should do so inside the initialize() function.
-	 * For instance change the value of gb.h to set the value of hubble parmaeter in your cosmology.
-	 * -----------------------------
-	 */
+{	 	
 
  	/**
- 	 * Initialize various ingredients for the rest of the code. More specifically, set the relavent directory paths, values of cosmological parameters and in general any parameter
- 	 * that has to be passed to CLASS Boltzmann code. Furthermore initialize some of the interpolators needed
- 	 * for the package.
+ 	 * Initialize various ingredients for the rest of the code. 
+ 	 * 
+ 	 * More specifically, initilize the filename and interpolator of SFR.
+ 	 * Read in the values of the parameters from .ini file and set the values of the elements of the glocal structure
+ 	 * Furthermore, the initialize() function returns a structure containing parameters such as as mimumum halo mass that are read from the .ini file.
  	 */
- 	initialize();
 
+	struct initialize_struct *init_struct  = (struct initialize_struct *)malloc(sizeof(struct initialize_struct));
 
-
- 	/**
-	 * -----------------------------
-	 * Change if you want to assume a different mass function or minimum halo mass
-	 * -----------------------------
-	 */
- 	
- 	/**
- 	 * Set the parameters for computing quantities related to spectral lines. 
- 	 * More specifically, the mimum mass for halos that can host line-emitting galaxies M_min
- 	 * and the the theoretical mass function to be used, ST: Sheth-Tormen, PSC: Press-Schecter, TK:: Tinker
- 	 */
-	double M_min   = 1.e9;
-	long mode_mf   = ST;
-
-
+	initialize(argv,init_struct);
 
  	/**
-	 * -----------------------------
-	 * Change if you want to limit the lines to include in your computation.
-	 * -----------------------------
-	 */
-	/**
-	 * Define which lines to compute, and how many points to have for z-interpolation
-	 * 
-	 * Note: The main line properties used in the calculation of the line power spectrum, which includes
+ 	 * Extract from the return struct of initialize() function, values of the parameters for computing quantities related to spectral lines. 
+ 	 * Note that these values were read from the .ini file within the initialize() function.
+ 	 * 
+ 	 * M_min : mimum mass for halos that can host line-emitting galaxies in unit of M_sun
+ 	 * mode_mf: switch to set the theoretical mass function to be used, ST: Sheth-Tormen, PSC: Press-Schecter, TK:: Tinker
+ 	 * ninterp: how many points to have for z-interpolation of line bias and first and second luminosity moments
+ 	 * nlines: the number of emission lines you want to include in the analysis
+ 	 * 
+ 	 * Note about ninterp: The main line properties used in the calculation of the line power spectrum, which includes
[…]
