+++ test.ini
@@ -1,0 +1,71 @@
+# values of cosmological parmaeters to pass to class
+# h:hubble 
+# Ocdm: Omega_cdm h^2
+# Ob: Omega_b h^2 
+# ns: primordial spectral index
+# As: primordial amplitude of fluctuations
+# kp: pivot point for primordial fluctuations
+h     =0.677      
+Ocdm = 0.11923
+Ob   = 0.02247
+ns   = 0.96824
+As   = 2.1085e-9
+kp   = 0.05
+
+#finger-of-god parameter, peculiar velocity dispersion
+sigFOG0 = 250.
+
+# Npars is the number of cosmological parmaters of the model 
+Npars = 6
+
+# PS_zmax: maximum redshift for matter power spectrum from class 
+# line_zmax: maximum redshift for interpolation of line quantities
+PS_zmax   = 6.
+line_zmax = 2.
+
+# values of min and max and number of points for the output of the mean, clusteringa dn shot components
+# for the test run, the values below dont matter since the mean and power are calculated for a single redshift and k-value
+mean_zmax   = 1.
+power_zmax  = 1.
+
+power_kmax  = 0.2
+power_kmin  = 1.e-3
+
+mean_nz     = 10
+power_nz    = 5
+power_nk    = 10
+
+# mode_mf sets the choice of halo mass function
+# Possible values: ST (Sheth-Tormen), 
+# PSC (Press-Schecter)
+# TK (Tinker)
+mode_mf = ST
+
+# Mh_min (in unit of M_sun) sets the minimum mass of teh halos that can host line-luminos galaxies
+Mh_min = 1.e9
+
+# ninterp is the number of interpolation points (in redshift) for calculating the line bias and first and second luminosity moments
+ninterp = 10
+
+# nlines is the number of emission lines you want to include in your calculation. The vlaue of nlines, can be at most 7. 
+nlines = 1
+
+#name of the lines to be included. You can pass at most 7 lines, CO10, CO21, CO32, CO43, CO54,CO65,CII. For example, if you want to only include CO10 and CII, you should set nlines = 2 and set line1 = CO10, line2 = CII, and the rest of the lines blank. The order of lines doesnt matter. For example in the above example you can set line1= CO10 and line2 = CII.
+line1 = CII
+line2 = 
+line3 = 
+line4 =
+line5 =
+line6 =
+line7 =
+
+# what to calculate. The "Tbar" option outputs the linear and quadratic line biases in addition to mean brightness temprature of the line.
+switch_Tbar    = Tbar
+switch_HMshot  = HMshot
+switch_HMclust = HMclust
+
+
+
+
+
+
