+++ miniminchia/tette.ini
@@ -1,13 +1,0 @@
-# Belle tette!
-Om = 3.232e-9 
-# Ma che figo
-
-non so che scrivere
-OL = 92.2 
-
-OpporcaVaccaMaiala = 1
-
-
-Macazzo = NonHaValoreNumerico
-
-suca!
