+++ miniminchia/makefile
@@ -1,2 +1,0 @@
-run.x: 
-	gcc -o run.x main.c
