+++ README.md
@@ -32,54 +32,57 @@ limHaloPT consists of 10 main modules, which include the following categories of
 - wnw_split.c: This module includes function to perform the splitting of matter power spectrum into its broadband and wiggle (BAO) contributions. 
 - survey_specs.c: This module includes functions to calculate quantities related to the choices of maximum and minnimum scales that can be probed by a given survey.
 - cosmology.c: This module includes function to compute the cosmological quantities at the level of background and perturbations. 
+- read_input.c: This module contains two functions initialize() and clean() which are the first and last functions called by the main() function. Initialize() reads in the input values from an .ini files.
+- setup_teardown.c: This module contains utilityy functions to read in an .ini file.
 - utilities.c: This module contains some utility functions, for example to build a dynamically allocated 1-dimensional array. These utility functions are used by the rest of the modules. <br>
 
 
 ## Compilation 
-- To compile, within the main directory of limHaloPT, type: make <br>
+- To compile, within the main directory of limHaloPT, type: "make" <br>
 This would create an executable called "limHaloPT" in the same directory, which you will use to run the code. If you modified the code, you need to first do "make clean" before doing "make". 
 
 The entire limHaloPT package was developed, compiled, and tested on Mac OS X, using gcc version 7.5.0 compiler. <br>
 
 
 ## Basic usage
-- To run the code, type  "./limHaloPT LCDM.ini"  
+- To run the code, in the main directory of limHaloPT, type  "./limHaloPT LCDM.ini"  
 
-Currently, the main output of limHaloPT are the mean brightness temprature, linear and quadratic biases, clustering and shot noise contributions of 6 emission spectral lines, depending on the switch that you set in the ini file. This computation of these functions are performed within main.c module. An example of the ini file is provided (LCDM.ini). If you want to call any of the functions of limHaloPT apart from those called by defacult, you neeed to add the function call to main.c, recompile the code by first cleaning the previous build using "make clean" and building teh package again with "make". 
+Currently, the main output of limHaloPT are the mean brightness temprature, linear and quadratic biases, clustering and shot noise contributions of 7 emission spectral lines, depending on the switch that you set in the ini file. The computation of these functions are performed within main.c module. An example of the ini file is provided (LCDM.ini). If you want to call any of the functions of limHaloPT, apart from those called by defacult, you neeed to add the function call to main.c, recompile the code by first cleaning the previous build using "make clean" and building the package again with "make". 
 
-Example: Lets say you want to compute the power spectrum of mean brightness temprature fluctuations for one or multiple emission lines, within halo-model, as a function of wavenumber, and at several redshifts. The relavent function that is called within main.c is PS_line_HM(). For this function call, first two arrays for values of redshifts and wavenumbers for which the power spectrum is computed are created. The number of emission lines to be included in the computation is set with "nlines" variable in the LCDM.ini file. Currently you can only include lines according to the order they appear You should also define for which emission lines do you want to compute the power spectrum. The choice of the lines is done in the first part of the main.c module, where the number of the lines, their name and their corresponding index is set. For example to compute the power spectrum for CO(1-0) and [CII] you would set 
[…]
