### 2022-04-06 d65b307
additional edits to readme
