+++ docs/latex/dir_74389ed8173ad57b461b9d623a1f3867.tex
@@ -1,7 +1,6 @@
 \hypertarget{dir_74389ed8173ad57b461b9d623a1f3867}{}\doxysubsection{Source Directory Reference}
 \label{dir_74389ed8173ad57b461b9d623a1f3867}\index{Source Directory Reference@{Source Directory Reference}}
-Directory dependency graph for Source\+:
-\nopagebreak
+Directory dependency graph for Source\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
