+++ docs/latex/d2/dbd/struct_cosmology.tex
@@ -11,8 +11,7 @@ Structure that holds varioud quantities that need to be evaluated for a given ch
 
 
 
-Collaboration diagram for Cosmology\+:
-\nopagebreak
+Collaboration diagram for Cosmology\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
