### 2022-04-06 amoradinejad/limHaloPT#11 @amoradinejad
I thank the referees for their helpful comments, which I believe have helped in improving limHaloPT package. To address the comment on the lack of examples of usage (#100, #11, #12) and also regarding making sure that all the requirements for JOSS publications [checklist] (https://joss.readthedocs.io/en/latest/review_checklist.html), I have made the following modifications to the code :

1) For ease of use, now the code uses an .ini file in which values of various parameters are set by the user. The .ini file includes descriptions of what is the significance of each parameter and what are the expected values. To read the  .ini file, two new (documented) modules are added to the package,  setup_teardown.c and  read_input.c.  The initilize() and clean() functions that were previously in main.c module are now moved to setup_teardown.c module. The initialize() function is modified to take input from .ini file.
2) The code now has three default outputs, values of mean brightness temperature (in addition to line biases), the clustering component of the line power spectrum, and the shot noise component of the line power spectrum.  In the .ini file, the user can set three switches determining whether or not to compute these three quantities.
3) By default the code saves the above outputs into files stored in "Output" folder.
4) I have modified the readme file to reflect the modifications to the code and include more detailed description of the component of the package. In particular, I have added a section titled "Structure of the package", which includes brief description of main modules of the package and their functionality.  Furthermore, there are now two sections describing "Compilation" and "Basic usage". Lastly, I added a section "Contributing" to address the JOSS requirement for "Community guidelines".
5) I have created a set of automated tests, to be found in "Test" subdirectory, which test the functionality of the main three functions which are called by default in main.c. The test can be run using the  test_LCDM.ini file, to be found in the same directory.

Thank you.
