### 2022-04-06 082531a
to address #14, README is edited to reflect how the test.c should be compiled and ran. Also discussion of usage of limHaloPT as an external library is added
