+++ docs/latex/index.tex
@@ -16,7 +16,23 @@ The source code of this package is publically avaialbel on Git\+Hub at \href{htt
 The lim\+Halo\+PT package calls various functions from \href{https://github.com/lesgourg/class_public}{\texttt{ CLASS}} Boltzman solver, including the matter power spectrum and transfer functions, growth factor etc. Therefore, you need to first download and compile CLASS code, create a \char`\"{} CLASS/lib/\char`\"{} folder and place the \char`\"{}libclass.\+a\char`\"{} in that folder. Furtehrmore, the loop calculations are performed with direct numerical integration, using routines of \href{http://www.feynarts.de/cuba/}{\texttt{ CUBA}} library. Lastly, the code heavily uses functions of \href{https://www.gnu.org/software/gsl/doc/html/}{\texttt{ GSL}} scientific library. Therfore, make sure that the two libraries are correctly linked to lim\+Halo\+PT by making necassary modifcations to the makefile (placed in Source directory) of lim\+Halo\+PT package.
 
 Note that in order to compute the luminosities of spectral lines, a model of star formation rate as a function of halo mass and redshift, SFR(\+M\+\_\+h,z), should be assumed. Currently, the implemented model uses SFR(\+M\+\_\+h,z) from \href{https://arxiv.org/abs/1207.6105}{\texttt{ Behroozi et al. (2013)}}. The necassary input file, \char`\"{}sfr\+\_\+release.\+dat\char`\"{}, is included in \char`\"{}\+Input/release-\/sfh\+\_\+z0\+\_\+z8\+\_\+052913/sfr/\char`\"{} subdirectory. ~\newline
- ~\hypertarget{index_autotoc_md4}{}\doxysubsubsection{Compilation}\label{index_autotoc_md4}
+ ~\hypertarget{index_autotoc_md4}{}\doxysubsubsection{Structure of the package}\label{index_autotoc_md4}
+lim\+Halo\+PT consists of 10 main modules, which include the following categories of functions\+:
+\begin{DoxyItemize}
+\item {\bfseries{\mbox{\hyperlink{main_8c}{main.\+c}}\+:}} This is the most external module, from which any function that you need is called. Depending on what quantities you want to calculate, you can modify the \mbox{\hyperlink{main_8c_a0ddf1224851353fc92bfbff6f499fa97}{main()}} function in \mbox{\hyperlink{main_8c}{main.\+c}} module (as marked in the code). Two example calls to functions which compute the clustering and shot noise contributions is included in \mbox{\hyperlink{main_8c}{main.\+c}} module. After adding the function calls in this module, you need to re-\/compile the code and then run it. Further details of the modules and descriptions of the functions can be found in documentation of the code.
+\item {\bfseries{\mbox{\hyperlink{ps__line__hm_8c}{ps\+\_\+line\+\_\+hm.\+c}}\+:}} This module includes various functions that are needed for computation of halo-\/model line power spectrum, including clustering and stochastic contributions beyond Poisson limit. At the moment, the redshift-\/space distrotions is not included in the halo model implementation.
[…]
