+++ README.md
@@ -94,7 +94,7 @@ Note that the test.c also can surve as an example of how you can link your code 
 
 
 ## Attribution
-You can use this package freely, provided that in your publication you cite the following paper; Moradinezhad, Nikakhtar, Keating, Castorina: [arXiv:2111.03717](https://arxiv.org/abs/2111.03717). Furthermore, since limHaloPT relies on CLASS Boltzman code, you should also cite at least this paper [arxiv:1104.2933](https://arxiv.org/abs/1104.2933) as required byy CLASS developers.<br>
+You can use this package freely, provided that in your publication you cite the following paper; Moradinezhad, Nikakhtar, Keating, Castorina: [arXiv:2111.03717](https://arxiv.org/abs/2111.03717). Furthermore, since limHaloPT relies on CLASS Boltzman code, you should also cite at least this paper [arxiv:1104.2933](https://arxiv.org/abs/1104.2933) as required by CLASS developers.<br>
 
 
 ## Contributing
