### 2022-04-11 871fbd7
to address #15, changed order of LIBFLAGS in makefile. Also created .gitkeep in Class/lib to include lib folder from git push/pull
