+++ docs/latex/dd/d24/survey__specs_8c.tex
@@ -6,8 +6,7 @@ Documented computation of some survey-\/related functions.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for survey\+\_\+specs.\+c\+:
-\nopagebreak
+Include dependency graph for survey\+\_\+specs.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -57,16 +56,14 @@ The function to be passed to Brent solver to compute kmax.
 {\em params} & Input\+: parameters to be passed to the solver \\
 \hline
 \end{DoxyParams}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{dd/d24/survey__specs_8c_a1eb59c068328f2b2a2cad533600c0d78_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -96,8 +93,7 @@ We follow Giannantonio et al. to for determining kmax, and use gsl Brent solver 
 kmax 
 \end{DoxyReturn}
 in short paper we used, 3.\+631872e-\/01 ; ~\newline
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -125,8 +121,7 @@ Compute the size of fundumental mode corresponding to the comoving volume enclos
 {\em fsky} & Input\+: sky-\/coverage of teh survey \\
 \hline
 \end{DoxyParams}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -155,16 +150,14 @@ Compute the comoving volume of a survey covering redshift up to z.
 \begin{DoxyReturn}{Returns}
 the comoving z-\/shell volume 
 \end{DoxyReturn}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=319pt]{dd/d24/survey__specs_8c_a91c4e0c62bcd3e179536fea62e8d53b5_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
