+++ docs/latex/d7/d10/line__ingredients_8c.tex
@@ -6,8 +6,7 @@ Documented line\+\_\+ingredients module.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for line\+\_\+ingredients.\+c\+:
-\nopagebreak
+Include dependency graph for line\+\_\+ingredients.\+c\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -133,16 +132,14 @@ un-\/normalized line bias ~\newline
 \end{DoxyReturn}
 In units of solar mass;
 
-In units of solar massHere is the call graph for this function\+:
-\nopagebreak
+In units of solar massHere is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d7/d10/line__ingredients_8c_aa050a054d3724145e3b46788bf279c79_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -172,16 +169,14 @@ The integrand function passed to hcubature integrator to compute the un-\/normal
 {\em fvalue} & Input\+: Array of values of the integrand of dimension fdim return the error status \\
 \hline
 \end{DoxyParams}
-Here is the call graph for this function\+:
-\nopagebreak
+Here is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d7/d10/line__ingredients_8c_a609c2a4b205e7f497d4b96657510fd5f_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -217,16 +212,14 @@ void ~\newline
 \end{DoxyReturn}
 Note that for PSC and ST mass functions, same form of the biases can be assumed, with different coefficents. See astro-\/ph/0006319
 
-Assuming spherical collapseHere is the call graph for this function\+:
-\nopagebreak
+Assuming spherical collapseHere is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d7/d10/line__ingredients_8c_a5e747f325f7196b6c042ddcfdfb17fab_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -262,16 +255,14 @@ This structure contains interpolators for computing the luminosity-\/weighted ma
 the total clustering line power spectrum, including the 1-\/ and 2-\/halo term ~\newline
  
 \end{DoxyReturn}
-CIIHere is the call graph for this function\+:
-\nopagebreak
+CIIHere is the call graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d7/d10/line__ingredients_8c_aa4696edc9829e9a3c2c1012a3e9d9dbe_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:
-\nopagebreak
+Here is the caller graph for this function\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
[…]
