+++ docs/latex/d6/da5/_global___structs_8h.tex
@@ -1,7 +1,6 @@
 \hypertarget{_global___structs_8h}{}\doxysubsection{Global\+\_\+\+Structs.\+h File Reference}
 \label{_global___structs_8h}\index{Global\_Structs.h@{Global\_Structs.h}}
-This graph shows which files directly or indirectly include this file\+:
-\nopagebreak
+This graph shows which files directly or indirectly include this file\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
