+++ docs/latex/df/dcb/header_8h.tex
@@ -37,16 +37,14 @@
 {\ttfamily \#include \char`\"{}ps\+\_\+line\+\_\+pt.\+h\char`\"{}}\newline
 {\ttfamily \#include \char`\"{}survey\+\_\+specs.\+h\char`\"{}}\newline
 {\ttfamily \#include \char`\"{}wnw\+\_\+split.\+h\char`\"{}}\newline
-Include dependency graph for header.\+h\+:
-\nopagebreak
+Include dependency graph for header.\+h\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{db/d0a/header_8h__incl}
 \end{center}
 \end{figure}
-This graph shows which files directly or indirectly include this file\+:
-\nopagebreak
+This graph shows which files directly or indirectly include this file\+:\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
