+++ docs/latex/d6/d32/cosmology_8c.tex
@@ -6,7 +6,8 @@ Documented cosmology module.
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for cosmology.\+c\+:\nopagebreak
+Include dependency graph for cosmology.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -152,7 +153,8 @@ luminosity distance d\+\_\+L = (1+z) d\+\_\+M angular diameter distance d\+\_\+A
 \begin{DoxyReturn}{Returns}
 D\+\_\+A 
 \end{DoxyReturn}
-junkHere is the caller graph for this function\+:\nopagebreak
+junkHere is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -177,7 +179,8 @@ Free the memory allocated to CLASS cosmology structure.
 \begin{DoxyReturn}{Returns}
 the error status 
 \end{DoxyReturn}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -206,7 +209,8 @@ Allocate memory and initialize the CLASS cosmology structure.
 \begin{DoxyReturn}{Returns}
 the error status 
 \end{DoxyReturn}
-pivot scale in unit of 1/\+MpcHere is the caller graph for this function\+:\nopagebreak
+pivot scale in unit of 1/\+MpcHere is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -234,7 +238,8 @@ Compute the comoving radial distance ~\newline
 \begin{DoxyReturn}{Returns}
 the double value D\+\_\+c 
 \end{DoxyReturn}
-junkHere is the caller graph for this function\+:\nopagebreak
+junkHere is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -261,7 +266,8 @@ Compute the cold dark matter concentration-\/mass relation.
 \begin{DoxyReturn}{Returns}
 the cdm concentration 
 \end{DoxyReturn}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -286,14 +292,16 @@ Free the memory allocated to cosmology structure.
 \begin{DoxyReturn}{Returns}
 the error status 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=310pt]{d6/d32/cosmology_8c_a8ddfed4a843beb762df124badc47011e_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -332,14 +340,16 @@ Allocate memory and initialize the cosmology structure, which includes the CLASS
 \begin{DoxyReturn}{Returns}
 an integer if succeeded 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
[…]
