+++ docs/latex/d1/dd5/ps__line__hm_8c.tex
@@ -6,7 +6,8 @@ Documented halo-\/model computation of line power spectrum, including clustering
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for ps\+\_\+line\+\_\+hm.\+c\+:\nopagebreak
+Include dependency graph for ps\+\_\+line\+\_\+hm.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -77,14 +78,16 @@ Compute the large-\/scale limit of P\+\_\+b2b2 loop.
 \begin{DoxyReturn}{Returns}
 b22\+\_\+ls 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d1/dd5/ps__line__hm_8c_aebed249f1a0afb0c6f7bcadcd3f1f4ee_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -114,14 +117,16 @@ The integrand function passed passed to Cuhre integration routine to compute lar
 {\em p} & Input\+: integration parmaeters return the error status \\
 \hline
 \end{DoxyParams}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d1/dd5/ps__line__hm_8c_a4e1bbff7f992103c9b1fa8f5abc1f408_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -172,14 +177,16 @@ Otherwise, NFW halo profile is assumed
 \begin{DoxyReturn}{Returns}
 void in unit of M\+\_\+sun/\+Mpc$^\wedge$3 
 \end{DoxyReturn}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d1/dd5/ps__line__hm_8c_a298eb563414c8e40e35481cb64c2f509_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
+Here is the caller graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -215,14 +222,16 @@ integrand of line 1halo term
 
 integrand of 2halo term proportional to b1, the linear local-\/in-\/matter halo bias
 
-integrand of 1halo term of line-\/matter cross-\/spectrumHere is the call graph for this function\+:\nopagebreak
+integrand of 1halo term of line-\/matter cross-\/spectrumHere is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
 \includegraphics[width=350pt]{d1/dd5/ps__line__hm_8c_ad72b2846b2e235de535da531f2ee30da_cgraph}
 \end{center}
 \end{figure}
-Here is the caller graph for this function\+:\nopagebreak
[…]
