+++ docs/latex/d0/d29/main_8c.tex
@@ -6,7 +6,8 @@ Documented main module, including functions to initilize and cleanup the cosmolo
 
 
 {\ttfamily \#include \char`\"{}header.\+h\char`\"{}}\newline
-Include dependency graph for main.\+c\+:\nopagebreak
+Include dependency graph for main.\+c\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode
@@ -62,7 +63,8 @@ Initialize the cosmology structure, which includes CLASS cosmology and \mbox{\hy
 
 \DoxyHorRuler{0}
  At the moment, lim\+Halo\+PT has three possible outputs; the mean brightness temprature, the clustering power or the shot power. The latter two are computed within halo model. The clustering power is real-\/space (RSD not included).\hypertarget{main.c_autotoc_md1}{}\doxyparagraph{If you want to call any other function of lim\+Halo\+PT, you should add them to to this section of the code}\label{main.c_autotoc_md1}
-Here is the call graph for this function\+:\nopagebreak
+Here is the call graph for this function\+:
+\nopagebreak
 \begin{figure}[H]
 \begin{center}
 \leavevmode

[…]
