### 2022-04-13 d0bc997
order of libs in makefile corrected to address #15
