### 2022-04-13 #4173 @eloisabentivegna
Hi @amoradinejad, do you have an estimate of when the restructuring will be complete?

@ziotom78, @abhimaniyar, is this currently a blocker for you?
