### 2022-04-14 #4173 @amoradinejad
@eloisabentivegna <https://github.com/eloisabentivegna> , the re-structuring of the code is completed, and I have addressed the comments made by the two referees. So on my side, all is done.

> On 13 Apr 2022, at 22:11, Eloisa Bentivegna ***@***.***> wrote:
>
>
> Hi @amoradinejad <https://github.com/amoradinejad>, do you have an estimate of when the restructuring will be complete?
>
> @ziotom78 <https://github.com/ziotom78>, @abhimaniyar <https://github.com/abhimaniyar>, is this currently a blocker for you?
>
