### 2022-05-07 #4173 @arfon
Looks like an author is missing an affiliation in the `paper.md`:

`  - name: Alberto Vallinotto`
