### 2022-05-07 #4173 @amoradinejad
The second author does not have an affiliation as he is not in academia. Do
all authors require an affiliation?

Thanks.

On Sat, May 7, 2022, 10:27 Eloisa Bentivegna ***@***.***>
wrote:

> Thanks!
>
> @amoradinejad <https://github.com/amoradinejad>, could you add the
> missing affiliation?
>
