### 2022-05-08 #4173 @eloisabentivegna
> Do all authors require an affiliation?

They do not. Please refer to the example at https://joss.readthedocs.io/en/latest/submitting.html?highlight=affiliation#example-paper-and-bibliography for how to list authors who do not have an affiliation.
