+++ Paper/paper.bib
@@ -1,11 +1,15 @@
-@article{MoradinezhadDizgah:2021vax,
+@article{MoradinezhadDizgah:2021dei,
     author = "Moradinezhad Dizgah, Azadeh and Nikakhtar, Farnik and Keating, Garrett K. and Castorina, Emanuele",
-    title = "{Precision Tests of CO and [CII] Power Spectra Models against Simulated Intensity Maps}",
+    title = "{Precision tests of CO and [CII] power spectra models against simulated intensity maps}",
     eprint = "2111.03717",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
-    month = "11",
-    year = "2021"
+    doi = "10.1088/1475-7516/2022/02/026",
+    journal = "JCAP",
+    volume = "02",
+    number = "02",
+    pages = "026",
+    year = "2022"
 }
 
 @article{Blas:2016sfa,
