### 2022-05-08 2e118f9
Dependencies section of joss paper updated
