+++ Paper/paper.md
@@ -48,16 +48,16 @@ Recently, there has been a shift in the cosmology community in publicly releasin
 
 # Dependencies
 
-The `limHaloPT` package calls various functions from [CLASS](https://github.com/lesgourg/class_public) Boltzman solver, including the matter power spectrum and transfer functions, growth factor etc. Therefore, prior to installation of `limHaloPT`, CLASS code should be compiled and the static library of "libclass.a" should be placed in Class/lib folder. Furtehrmore, the loop calculations are performed with direct numerical integration, using routines of [CUBA](http://www.feynarts.de/cuba/) library. Lastly, the code heavily uses functions of [GSL](https://www.gnu.org/software/gsl/doc/html/)scientific library. Therfore, the two libraries should be correctly linked to `limHaloPT` by making necassary modifcations to the makefile (placed in the main directory). 
+The `limHaloPT` package calls various functions from [CLASS](https://github.com/lesgourg/class_public) Boltzman solver, including the matter power spectrum and transfer functions, growth factor etc. Therefore, prior to installation of `limHaloPT`, CLASS code should be compiled and the static library of "libclass.a" should be placed in "Class/lib" folder. Furtehrmore, the loop calculations are performed with direct numerical integration, using routines of [CUBA](http://www.feynarts.de/cuba/) library. Lastly, the code heavily uses functions of [GSL](https://www.gnu.org/software/gsl/doc/html/)scientific library. Therfore, the two libraries should be correctly linked to `limHaloPT` by making necassary modifcations to the makefile (placed in the main directory). 
 
 In order to compute the luminosities of spectral lines, a model of star formation rate as a function of halo mass and redshift, SFR(M_h,z), should be assumed. Currently, the implemented model uses SFR(M_h,z) from [Behroozi et al. (2013)](https://arxiv.org/abs/1207.6105). The necassary input file, "sfr_release.dat", is included in "Input/release-sfh_z0_z8_052913/sfr/" subdirectory. 
 
 
 # Usage
 
-Currently, the main output of `limHaloPT` are the mean brightness temprature, linear and quadratic biases, clustering and shot noise contributions of seven emission spectral lines. The lines to compute are set using a switch in the ini file, an example of the ini file is provided in the package(LCDM.ini). More detailed description of the main outputs is given in the code documentation on github reporistory. By default, in addition to saving the output for mean brightness temprature, biases, shot and clustering components of power spectrum, when computing the clustering and shot powers, individual loop contributions (in the former) and individual beyond-Possion contribution to the shot (in the latter), are also saved to output files which are stored in "Output" directory. 
+Currently, the main output of `limHaloPT` are the mean brightness temprature, linear and quadratic biases, clustering and shot noise contributions of seven emission spectral lines. The lines to compute are set using a switch in the ini file, an example of the ini file is provided in the package (see "LCDM.ini"). More detailed description of the main outputs is given in the code documentation on github reporistory. By default, in addition to saving the output for mean brightness temprature, biases, shot and clustering components of power spectrum, when computing the clustering and shot powers, individual loop contributions (in the former) and individual beyond-Possion contribution to the shot (in the latter), are also saved to output files which are stored in "Output" directory. 
 
-In addition to using `limHaloPT` through "main.c" module which calls three specific functions, the package can also be exported as an external library to any other C code. Upon compilation of the package, a static library called `liblimHaloPT.a` is created and placed in `lib` directory, which can then be linked to an external C code. Example of how to link an external code to liblimHaloPT.a static library can be found in "Test" directory.
[…]
