### 2022-05-09 165faba
typos in Joss paper corrected
