### 2022-05-12 #4173 @amoradinejad
@eloisabentivegna <https://github.com/eloisabentivegna> thanks for the info, I updated the paper with the affiliation as described in the link.

> On 8 May 2022, at 23:34, Eloisa Bentivegna ***@***.***> wrote:
>
>
> Do all authors require an affiliation?
>
> They do not. Please refer to the example at https://joss.readthedocs.io/en/latest/submitting.html?highlight=affiliation#example-paper-and-bibliography <https://joss.readthedocs.io/en/latest/submitting.html?highlight=affiliation#example-paper-and-bibliography> for how to list authors who do not have an affiliation.
>
