### 2022-05-19 #4173 @amoradinejad
The article looks good and I don’t have any correction, apart from the fact that I would like to have my full last name at the citation in the footnote. So please replace “Dizgah et al.” with “Moradinezhad Dizgah et al”.

Best Regards,
Azadeh Moradinezhad

> On 17 May 2022, at 10:15, The Open Journals editorial robot ***@***.***> wrote:
>
>
> 👉📄 Download article proof <https://raw.githubusercontent.com/openjournals/joss-papers/joss.04173/joss.04173/10.21105.joss.04173.pdf> 📄 View article proof on GitHub <https://github.com/openjournals/joss-papers/blob/joss.04173/joss.04173/10.21105.joss.04173.pdf> 📄 👈
>
