+++ Paper/paper.md
@@ -32,32 +32,32 @@ bibliography: paper.bib
 
 # Summary
 
-`limHaloPT` is a modular numerical package, written in C, for computing the clustering and shot-noise contributions to the power spectrum of line intensity/temperature fluctuations using the halo-model framework. This package is the first publicly available code, which combines the one-loop prediction of the halo power spectrum and the halo-model framework to model the power spectrum of emission lines originating from star-forming galaxies. Furthermore, the code includes routines to compute the stochastic contributions to the line power spectrum beyond the Poisson approximation. Several utility functions, e.g., for computing the theoretical halo mass functions, halo biases, one-loop halo power spectrum, are provided in the package, which can be used in contexts other than Line intensity mapping (LIM). This code is released together with a scientific publication [@MoradinezhadDizgah:2021dei], in which details of the implemented model and the comparison of model predictions against simulated intensity maps are presented. The current version of the code includes the first six rotational ladder of carbon monoxide, CO, and fine structure line of ionized carbon [CII], and the model of the power spectrum is limited to real space.
+`limHaloPT` is a modular numerical package, written in C, for computing the clustering and shot-noise contributions to the power spectrum of line intensity/temperature fluctuations using the halo-model framework. This package is the first publicly available code, which combines the one-loop prediction of the halo power spectrum and the halo-model framework to model the power spectrum of emission lines originating from star-forming galaxies. Furthermore, the code includes routines to compute the stochastic contributions to the line power spectrum beyond the Poisson approximation. Several utility functions, e.g., for computing the theoretical halo mass functions, halo biases, one-loop halo power spectrum, are provided in the package, which can be used in contexts other than Line Intensity Mapping (LIM). This code is released together with a scientific publication [@MoradinezhadDizgah:2021dei], in which details of the implemented model and the comparison of model predictions against simulated intensity maps are presented. The current version of the code includes the first six rotational ladders of carbon monoxide, CO, and fine structure line of ionized carbon [CII], and the model of the power spectrum is limited to real space.
 
 
 # Scientific Context and Statement of Need
 
-Line intensity mapping is a novel technique to map the large-scale structure (LSS) of the Universe by measuring aggregate emission of the atomic and molecular emission lines from the unresolved sources [@Kovetz:2017agg]. Measurements of spatial fluctuations and frequency of the line provide a 3-dimensional map of the LSS, whose statistical properties capture a significant amount of information about astrophysics and cosmology. To fully exploit this rich data, accurate theoretical models of the signal and efficient numerical codes for evaluating the models are crucial. 
+LIM is a novel technique to map the large-scale structure (LSS) of the Universe by measuring the aggregate intensity of the atomic and molecular emission lines from the unresolved sources [@Kovetz:2017agg]. Measurements of spatial fluctuations and frequency of the line provide a 3-dimensional map of the LSS, whose statistical properties capture a significant amount of information about astrophysics and cosmology. To fully exploit this rich data, accurate theoretical models of the signal and efficient numerical codes for evaluating the models are crucial.
 
[…]
