+++ Paper/paper.bib
@@ -101,7 +101,8 @@
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
     month = "11",
-    year = "2021"
+    year = "2021",
+    doi = "10.1088/1475-7516/2022/02/026"
 }
 
 @article{Schaan:2021hhy,
