### 2022-05-20 #4173 @eloisabentivegna
@amoradinejad, I am not sure how to fix the footer problem you mention, but I am looking into it. In the meantime, I have created a pull request to your repository with some minor (mostly cosmetic) suggestions. Please take a look and integrate, if you approve.
