### 2022-05-20 #4173 @eloisabentivegna
> I would like to have my full last name at the citation in the footnote

@amoradinejad, there are some instructions about fine-graining author names at https://github.com/openjournals/inara/blob/main/docs/names.md. Could you take a look and try the fix?
