+++ Paper/paper.md
@@ -5,7 +5,8 @@ tags:
   - cosmology
   - intensity mapping
 authors:
-  - name: Azadeh Moradinezhad Dizgah
+  - given names: Azadeh 
+  - surname: Moradinezhad Dizgah
     orcid: 0000-0001-8841-9989
     affiliation: 1
   - name: Alberto Vallinotto  
