+++ Paper/paper.md
@@ -5,10 +5,11 @@ tags:
   - cosmology
   - intensity mapping
 authors:
-  - given names: Azadeh 
-  - surname: Moradinezhad Dizgah
-    orcid: 0000-0001-8841-9989
-    affiliation: 1
+  - name:
+        given names: Azadeh 
+        surname: Moradinezhad Dizgah
+        orcid: 0000-0001-8841-9989
+        affiliation: 1
   - name: Alberto Vallinotto  
     affiliation: 2
   - name: Farnik Nikakhtar
