### 2022-05-27 #4173 @amoradinejad
@eloisabentivegna <https://github.com/eloisabentivegna> , sorry for the missing information.

Here is the checklist:
 Make a tagged release of your software, and list the version tag of the archived version here: Version v2.0.0 <https://zenodo.org/record/6576183>
 Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository): ARCHIVED IN ZENODO
 Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.  CORRECTED THE TITLE ON ZENODO
 Please list the DOI of the archived version here. DOI: 10.5281/zenodo.6576183

> On 25 May 2022, at 00:41, Eloisa Bentivegna ***@***.***> wrote:
>
>
> Thanks @amoradinejad <https://github.com/amoradinejad>. There seems to be an error with the compilation, due to your affiliation this time -- let me look into it.
>
> In the meantime, could you please go through the checklist above? Items 1 and 3 have not been addressed (we're missing the version tag, and the title on Zenodo does not match the submission title).
>
