### 2022-07-07 296ada8
case fixes in bib
