+++ Paper/paper.bib
@@ -1,6 +1,6 @@
 @article{MoradinezhadDizgah:2021dei,
     author = "Moradinezhad Dizgah, Azadeh and Nikakhtar, Farnik and Keating, Garrett K. and Castorina, Emanuele",
-    title = "{Precision tests of CO and [CII] power spectra models against simulated intensity maps}",
+    title = {Precision tests of {CO} and [{CII}] power spectra models against simulated intensity maps},
     eprint = "2111.03717",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
@@ -14,7 +14,7 @@
 
 @article{Blas:2016sfa,
     author = "Blas, Diego and Garny, Mathias and Ivanov, Mikhail M. and Sibiryakov, Sergey",
-    title = "{Time-Sliced Perturbation Theory II: Baryon Acoustic Oscillations and Infrared Resummation}",
+    title = {Time-Sliced Perturbation Theory {II}: Baryon Acoustic Oscillations and Infrared Resummation},
     eprint = "1605.02149",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
@@ -28,7 +28,7 @@
 
 @article{Senatore:2014via,
     author = "Senatore, Leonardo and Zaldarriaga, Matias",
-    title = "{The IR-resummed Effective Field Theory of Large Scale Structures}",
+    title = {The {IR}-resummed Effective Field Theory of Large Scale Structures},
     eprint = "1404.5954",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
@@ -41,7 +41,7 @@
 
 @article{Senatore:2017pbn,
     author = "Senatore, Leonardo and Trevisan, Gabriele",
-    title = "{On the IR-Resummation in the EFTofLSS}",
+    title = {On the {IR}-Resummation in the {EFTofLSS}},
     eprint = "1710.02178",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
@@ -54,7 +54,7 @@
 
 @article{Zuntz:2014csq,
     author = "Zuntz, Joe and Paterno, Marc and Jennings, Elise and Rudd, Douglas and Manzotti, Alessandro and Dodelson, Scott and Bridle, Sarah and Sehrish, Saba and Kowalkowski, James",
-    title = "{CosmoSIS: modular cosmological parameter estimation}",
+    title = {{CosmoSIS}: modular cosmological parameter estimation},
     eprint = "1409.3409",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
@@ -67,7 +67,7 @@
 }
 
 @article{Blas2011,
-   title={The Cosmic Linear Anisotropy Solving System (CLASS). Part II: Approximation schemes},
+   title={The Cosmic Linear Anisotropy Solving System ({CLASS}). Part {II}: Approximation schemes},
    volume={2011},
    ISSN={1475-7516},
    url={http://dx.doi.org/10.1088/1475-7516/2011/07/034},
@@ -82,7 +82,7 @@
 }
 @article{Hand:2017pqn,
     author = "Hand, Nick and Feng, Yu and Beutler, Florian and Li, Yin and Modi, Chirag and Seljak, Uros and Slepian, Zachary",
-    title = "{nbodykit: an open-source, massively parallel toolkit for large-scale structure}",
+    title = {{nbodykit}: an open-source, massively parallel toolkit for large-scale structure},
     eprint = "1712.05834",
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.IM",
@@ -96,7 +96,7 @@
 
 @article{MoradinezhadDizgah:2021,
[…]
