### 2022-07-09 62522d6
Journal titles updated.
