+++ Paper/paper.bib
@@ -5,7 +5,7 @@
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
     doi = "10.1088/1475-7516/2022/02/026",
-    journal = "JCAP",
+    journal = "Journal of Cosmology and Astroparticle Physics",
     volume = "02",
     number = "02",
     pages = "026",
@@ -20,7 +20,7 @@
     primaryClass = "astro-ph.CO",
     reportNumber = "CERN-TH-2016-059, INR-TH-2016-009",
     doi = "10.1088/1475-7516/2016/07/028",
-    journal = "JCAP",
+    journal = "Journal of Cosmology and Astroparticle Physics",
     volume = "07",
     pages = "028",
     year = "2016"
@@ -33,7 +33,7 @@
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
     doi = "10.1088/1475-7516/2015/02/013",
-    journal = "JCAP",
+    journal = "Journal of Cosmology and Astroparticle Physics",
     volume = "02",
     pages = "013",
     year = "2015"
@@ -46,7 +46,7 @@
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
     doi = "10.1088/1475-7516/2018/05/019",
-    journal = "JCAP",
+    journal = "Journal of Cosmology and Astroparticle Physics",
     volume = "05",
     pages = "019",
     year = "2018"
@@ -60,7 +60,7 @@
     primaryClass = "astro-ph.CO",
     reportNumber = "FERMILAB-PUB-14-408-A",
     doi = "10.1016/j.ascom.2015.05.005",
-    journal = "Astron. Comput.",
+    journal = "Astronomy and Computing",
     volume = "12",
     pages = "45--59",
     year = "2015"
@@ -87,22 +87,11 @@
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.IM",
     doi = "10.3847/1538-3881/aadae0",
-    journal = "Astron. J.",
+    journal = "The Astronomical Journal",
     volume = "156",
     number = "4",
     pages = "160",
     year = "2018"
-}
-
-@article{MoradinezhadDizgah:2021,
-    author = "Moradinezhad Dizgah, Azadeh and Nickakhtar, Farnik and Keating, Garret K. and Castorina, Emanuele ",
-    title = {Precision Tests of {CO} and [{CII}] Power SpectraModels against Simulated Intensity Maps},
-    eprint = "2111.XXXXX",
-    archivePrefix = "arXiv",
-    primaryClass = "astro-ph.CO",
-    month = "11",
-    year = "2021",
-    doi = "10.1088/1475-7516/2022/02/026"
 }
 
 @article{Schaan:2021hhy,
@@ -112,7 +101,7 @@
     archivePrefix = "arXiv",
     primaryClass = "astro-ph.CO",
     doi = "10.1088/1475-7516/2021/05/067",
-    journal = "JCAP",
+    journal = "Journal of Cosmology and Astroparticle Physics",
     volume = "05",
     pages = "067",
     year = "2021"
@@ -134,7 +123,7 @@
     eprint = "astro-ph/0001493",
     archivePrefix = "arXiv",
     doi = "10.1046/j.1365-8711.2000.03715.x",
-    journal = "Mon. Not. Roy. Astron. Soc.",
+    journal = "Monthly Notices of the Royal Astronomical Society",
     volume = "318",
     pages = "203",
     year = "2000"
@@ -147,7 +136,7 @@
     archivePrefix = "arXiv",
     reportNumber = "FERMILAB-PUB-02-284-A",
     doi = "10.1016/S0370-1573(02)00276-4",
-    journal = "Phys. Rept.",
[…]
