### 2022-07-09 #4173 @amoradinejad
Yes, the two entries are the same. One is now removed.

> On 7 Jul 2022, at 21:25, Daniel S. Katz ***@***.***> wrote:
>
>
> Also, are bib entries MoradinezhadDizgah:2021dei and oradinezhadDizgah:2021 the same? If so, can one be removed?
>
