### 2022-07-09 #4173 @amoradinejad
I accepted the merge request. All looks good to me. I updated the journal name to have full names.

> On 7 Jul 2022, at 21:24, Daniel S. Katz ***@***.***> wrote:
>
>
> 👋 I'm the AEiC this week, handing the final processing of this submission. I've suggested some changes in amoradinejad/limHaloPT#18 <https://github.com/amoradinejad/limHaloPT/pull/18> - please merge, or let me know what you disagree with. In addition, after merging, please update the journals in your bib entries to not be abbreviations such as JCAP, JHEP, etc. since JOSS is a general journal where these abbreviations are not likely to be known by all readers.
>
