### 2022-07-09 #4173 @amoradinejad
Merge request accepted.

> On 9 Jul 2022, at 19:28, Arfon Smith ***@***.***> wrote:
>
>
> @amoradinejad <https://github.com/amoradinejad> – could you please merge this PR which fixes an issue with the first author name? amoradinejad/limHaloPT#19 <https://github.com/amoradinejad/limHaloPT/pull/19>
