### 2022-07-09 #4173 @arfon
@xuanxu @tarleb – there was an error when publishing this paper (when the author name was written as follows: https://raw.githubusercontent.com/amoradinejad/limHaloPT/262d3fae1c5478e107491b0d462300e7ee107657/Paper/paper.md)

This resulted in the following output on the JOSS site:

I proposed a change which I thought would fix this (see [paper.md](https://raw.githubusercontent.com/amoradinejad/limHaloPT/master/Paper/paper.md)) based on the [Inara docs](
https://github.com/openjournals/inara/blob/main/docs/names.md):

```
authors:
  - given-names: Azadeh
    surname: Moradinezhad Dizgah
    orcid: 0000-0001-8841-9989
    affiliation: 1
  - name: Alberto Vallinotto
    affiliation: 2
  - name: Farnik Nikakhtar
    orcid: 0000-0002-3641-4366
    affiliation: 3
```

but it doesn't seem to have worked. Could you advise what the right formatting of the metadata is here?
