### 2022-07-10 #4173 @xuanxu
I think the error happens in the Ruby code parsing names, currently is not accepting the same keys as Inara.
I'll fix it and release a new version of the Theoj gem.
